Refactor: split off ESIL from analysis_arm_cs.c
