+++ librz/analysis/p/analysis_arm_cs.c
@@ -10,58 +10,8 @@
 #include <rz_util/rz_assert.h>
 #include "./analysis_arm_hacks.inc"
 
-/* arm64 */
-#define IMM64(x)   (ut64)(insn->detail->arm64.operands[x].imm)
-#define INSOP64(x) insn->detail->arm64.operands[x]
-
-/* arm32 */
-#define REG(x)       rz_str_get_null(cs_reg_name(*handle, insn->detail->arm.operands[x].reg))
-#define REG64(x)     rz_str_get_null(cs_reg_name(*handle, insn->detail->arm64.operands[x].reg))
-#define REGID64(x)   insn->detail->arm64.operands[x].reg
-#define REGID(x)     insn->detail->arm.operands[x].reg
-#define IMM(x)       (ut32)(insn->detail->arm.operands[x].imm)
-#define INSOP(x)     insn->detail->arm.operands[x]
-#define MEMBASE(x)   rz_str_get_null(cs_reg_name(*handle, insn->detail->arm.operands[x].mem.base))
-#define MEMBASE64(x) rz_str_get_null(cs_reg_name(*handle, insn->detail->arm64.operands[x].mem.base))
-#define REGBASE(x)   insn->detail->arm.operands[x].mem.base
-#define REGBASE64(x) insn->detail->arm64.operands[x].mem.base
-// s/index/base|reg/
-#define MEMINDEX(x)      rz_str_get_null(cs_reg_name(*handle, insn->detail->arm.operands[x].mem.index))
-#define HASMEMINDEX(x)   (insn->detail->arm.operands[x].mem.index != ARM_REG_INVALID)
-#define MEMINDEX64(x)    rz_str_get_null(cs_reg_name(*handle, insn->detail->arm64.operands[x].mem.index))
-#define HASMEMINDEX64(x) (insn->detail->arm64.operands[x].mem.index != ARM64_REG_INVALID)
-#define ISMEMINDEXSUB(x) insn->detail->arm.operands[x].subtracted
-#define MEMDISP(x)       insn->detail->arm.operands[x].mem.disp
-#define MEMDISP64(x)     (ut64) insn->detail->arm64.operands[x].mem.disp
-#define ISIMM(x)         (insn->detail->arm.operands[x].type == ARM_OP_IMM)
-#define ISIMM64(x)       (insn->detail->arm64.operands[x].type == ARM64_OP_IMM)
-#define ISREG(x)         (insn->detail->arm.operands[x].type == ARM_OP_REG)
-#define ISREG64(x)       (insn->detail->arm64.operands[x].type == ARM64_OP_REG)
-#define ISMEM(x)         (insn->detail->arm.operands[x].type == ARM_OP_MEM)
-#define ISMEM64(x)       (insn->detail->arm64.operands[x].type == ARM64_OP_MEM)
-#define EXT64(x)         decode_sign_ext(insn->detail->arm64.operands[x].ext)
-
-#if CS_API_MAJOR > 3
-#define LSHIFT(x)     insn->detail->arm.operands[x].mem.lshift
-#define LSHIFT2(x)    insn->detail->arm.operands[x].shift.value // Dangerous, returns value even if isn't LSL
-#define LSHIFT2_64(x) insn->detail->arm64.operands[x].shift.value
-#else
-#define LSHIFT(x)     0
-#define LSHIFT2(x)    0
-#define LSHIFT2_64(x) 0
-#endif
-#define OPCOUNT()     insn->detail->arm.op_count
-#define OPCOUNT64()   insn->detail->arm64.op_count
-#define ISSHIFTED(x)  (insn->detail->arm.operands[x].shift.type != ARM_SFT_INVALID && insn->detail->arm.operands[x].shift.value != 0)
-#define SHIFTTYPE(x)  insn->detail->arm.operands[x].shift.type
-#define SHIFTVALUE(x) insn->detail->arm.operands[x].shift.value
-
-#define ISWRITEBACK32() insn->detail->arm.writeback
-#define ISPREINDEX32()  (((OPCOUNT() == 2) && (ISMEM(1)) && (ISWRITEBACK32())) || ((OPCOUNT() == 3) && (ISMEM(2)) && (ISWRITEBACK32())))
-#define ISPOSTINDEX32() (((OPCOUNT() == 3) && (ISIMM(2) || ISREG(2)) && (ISWRITEBACK32())) || ((OPCOUNT() == 4) && (ISIMM(3) || ISREG(3)) && (ISWRITEBACK32())))
-#define ISWRITEBACK64() (insn->detail->arm64.writeback == true)
-#define ISPREINDEX64()  (((OPCOUNT64() == 2) && (ISMEM64(1)) && (ISWRITEBACK64())) || ((OPCOUNT64() == 3) && (ISMEM64(2)) && (ISWRITEBACK64())))
-#define ISPOSTINDEX64() (((OPCOUNT64() == 3) && (ISIMM64(2)) && (ISWRITEBACK64())) || ((OPCOUNT64() == 4) && (ISIMM64(3)) && (ISWRITEBACK64())))
+#include "../arch/arm/arm_cs.h"
+#include "../arch/arm/arm_accessors.h"
 
 typedef struct arm_cs_context_t {
 	HtUU *ht_itblock;
@@ -70,20 +20,6 @@ typedef struct arm_cs_context_t {
 	int omode;
 	int obits;
 } ArmCSContext;
-
-static const ut64 bitmask_by_width[] = {
-	0x1, 0x3, 0x7, 0xf, 0x1f, 0x3f, 0x7f, 0xff, 0x1ff, 0x3ff, 0x7ff,
-	0xfff, 0x1fff, 0x3fff, 0x7fff, 0xffff, 0x1ffff, 0x3ffff, 0x7ffff,
-	0xfffff, 0x1fffff, 0x3fffff, 0x7fffff, 0xffffff, 0x1ffffffLL, 0x3ffffffLL,
-	0x7ffffffLL, 0xfffffffLL, 0x1fffffffLL, 0x3fffffffLL, 0x7fffffffLL, 0xffffffffLL,
-	0x1ffffffffLL, 0x3ffffffffLL, 0x7ffffffffLL, 0xfffffffffLL, 0x1fffffffffLL,
-	0x3fffffffffLL, 0x7fffffffffLL, 0xffffffffffLL, 0x1ffffffffffLL, 0x3ffffffffffLL,
-	0x7ffffffffffLL, 0xfffffffffffLL, 0x1fffffffffffLL, 0x3fffffffffffLL, 0x7fffffffffffLL,
-	0xffffffffffffLL, 0x1ffffffffffffLL, 0x3ffffffffffffLL, 0x7ffffffffffffLL,
-	0xfffffffffffffLL, 0x1fffffffffffffLL, 0x3fffffffffffffLL, 0x7fffffffffffffLL,
-	0xffffffffffffffLL, 0x1ffffffffffffffLL, 0x3ffffffffffffffLL, 0x7ffffffffffffffLL,
-	0xfffffffffffffffLL, 0x1fffffffffffffffLL, 0x3fffffffffffffffLL, 0x7fffffffffffffffLL, 0xffffffffffffffffLL
-};
 
 static const char *shift_type_name(arm_shifter type) {
 	switch (type) {
@@ -365,47 +301,6 @@ static void opex(RzStrBuf *buf, csh handle, cs_insn *insn) {
 	pj_free(pj);
 }
 
-static int arm64_reg_width(int reg) {
-	switch (reg) {
-	case ARM64_REG_W0:
-	case ARM64_REG_W1:
-	case ARM64_REG_W2:
-	case ARM64_REG_W3:
-	case ARM64_REG_W4:
-	case ARM64_REG_W5:
-	case ARM64_REG_W6:
-	case ARM64_REG_W7:
-	case ARM64_REG_W8:
-	case ARM64_REG_W9:
-	case ARM64_REG_W10:
-	case ARM64_REG_W11:
-	case ARM64_REG_W12:
-	case ARM64_REG_W13:
-	case ARM64_REG_W14:
-	case ARM64_REG_W15:
-	case ARM64_REG_W16:
-	case ARM64_REG_W17:
-	case ARM64_REG_W18:
-	case ARM64_REG_W19:
-	case ARM64_REG_W20:
-	case ARM64_REG_W21:
-	case ARM64_REG_W22:
-	case ARM64_REG_W23:
-	case ARM64_REG_W24:
-	case ARM64_REG_W25:
-	case ARM64_REG_W26:
-	case ARM64_REG_W27:
-	case ARM64_REG_W28:
-	case ARM64_REG_W29:
-	case ARM64_REG_W30:
-		return 32;
-		break;
-	default:
-		break;
-	}
-	return 64;
-}
-
 static const char *cc_name64(arm64_cc cc) {
 	switch (cc) {
 	case ARM64_CC_EQ: // Equal
@@ -652,2035 +547,6 @@ static void opex64(RzStrBuf *buf, csh handle, cs_insn *insn) {
 	rz_strbuf_init(buf);
 	rz_strbuf_append(buf, pj_string(pj));
 	pj_free(pj);
-}
-
-static int decode_sign_ext(arm64_extender extender) {
-	switch (extender) {
-	case ARM64_EXT_UXTB:
-	case ARM64_EXT_UXTH:
-	case ARM64_EXT_UXTW:
-	case ARM64_EXT_UXTX:
-		return 0; // nothing needs to be done for unsigned
-	case ARM64_EXT_SXTB:
-		return 8;
-	case ARM64_EXT_SXTH:
-		return 16;
-	case ARM64_EXT_SXTW:
-		return 32;
-	case ARM64_EXT_SXTX:
-		return 64;
-	default:
-		break;
-	}
-
-	return 0;
-}
-
-static const char *decode_shift(arm_shifter shift) {
-	const char *E_OP_SR = ">>";
-	const char *E_OP_SL = "<<";
-	const char *E_OP_RR = ">>>";
-	const char *E_OP_ASR = ">>>>";
-	const char *E_OP_VOID = "";
-
-	switch (shift) {
-	case ARM_SFT_ASR:
-	case ARM_SFT_ASR_REG:
-		return E_OP_ASR;
-
-	case ARM_SFT_LSR:
-	case ARM_SFT_LSR_REG:
-		return E_OP_SR;
-
-	case ARM_SFT_LSL:
-	case ARM_SFT_LSL_REG:
-		return E_OP_SL;
-
-	case ARM_SFT_ROR:
-	case ARM_SFT_RRX:
-	case ARM_SFT_ROR_REG:
-	case ARM_SFT_RRX_REG:
-		return E_OP_RR;
-
-	default:
-		break;
-	}
-	return E_OP_VOID;
-}
-
-static const char *decode_shift_64(arm64_shifter shift) {
-	const char *E_OP_SR = ">>";
-	const char *E_OP_SL = "<<";
-	const char *E_OP_RR = ">>>";
-	const char *E_OP_VOID = "";
-
-	switch (shift) {
-	case ARM64_SFT_ASR:
-	case ARM64_SFT_LSR:
-		return E_OP_SR;
-
-	case ARM64_SFT_LSL:
-	case ARM64_SFT_MSL:
-		return E_OP_SL;
-
-	case ARM64_SFT_ROR:
-		return E_OP_RR;
-
-	default:
-		break;
-	}
-	return E_OP_VOID;
-}
-
-#define DECODE_SHIFT(x)   decode_shift(insn->detail->arm.operands[x].shift.type)
-#define DECODE_SHIFT64(x) decode_shift_64(insn->detail->arm64.operands[x].shift.type)
-
-static unsigned int regsize32(cs_insn *insn, int n) {
-	rz_return_val_if_fail(n >= 0 && n < insn->detail->arm.op_count, 0);
-	unsigned int reg = insn->detail->arm.operands[n].reg;
-	if (reg >= ARM_REG_D0 && reg <= ARM_REG_D31) {
-		return 8;
-	}
-	if (reg >= ARM_REG_Q0 && reg <= ARM_REG_Q15) {
-		return 16;
-	}
-	return 4; // s0-s31, r0-r15
-}
-
-static int regsize64(cs_insn *insn, int n) {
-	unsigned int reg = insn->detail->arm64.operands[n].reg;
-	if ((reg >= ARM64_REG_S0 && reg <= ARM64_REG_S31) ||
-		(reg >= ARM64_REG_W0 && reg <= ARM64_REG_W30) ||
-		reg == ARM64_REG_WZR) {
-		return 4;
-	}
-	if (reg >= ARM64_REG_B0 && reg <= ARM64_REG_B31) {
-		return 1;
-	}
-	if (reg >= ARM64_REG_H0 && reg <= ARM64_REG_H31) {
-		return 2;
-	}
-	if (reg >= ARM64_REG_Q0 && reg <= ARM64_REG_Q31) {
-		return 16;
-	}
-	return 8;
-}
-
-#define REGSIZE64(x) regsize64(insn, x)
-#define REGSIZE32(x) regsize32(insn, x)
-
-// return postfix
-const char *arm_prefix_cond(RzAnalysisOp *op, int cond_type) {
-	const char *close_cond[2];
-	close_cond[0] = "\0";
-	close_cond[1] = ",}\0";
-	int close_type = 0;
-	switch (cond_type) {
-	case ARM_CC_EQ:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "zf,?{,");
-		break;
-	case ARM_CC_NE:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "zf,!,?{,");
-		break;
-	case ARM_CC_HS:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "cf,?{,");
-		break;
-	case ARM_CC_LO:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "cf,!,?{,");
-		break;
-	case ARM_CC_MI:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "nf,?{,");
-		break;
-	case ARM_CC_PL:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "nf,!,?{,");
-		break;
-	case ARM_CC_VS:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "vf,?{,");
-		break;
-	case ARM_CC_VC:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "vf,!,?{,");
-		break;
-	case ARM_CC_HI:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "cf,zf,!,&,?{,");
-		break;
-	case ARM_CC_LS:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "cf,!,zf,|,?{,");
-		break;
-	case ARM_CC_GE:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "nf,vf,^,!,?{,");
-		break;
-	case ARM_CC_LT:
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "nf,vf,^,?{,");
-		break;
-	case ARM_CC_GT:
-		// zf == 0 && nf == vf
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "zf,!,nf,vf,^,!,&,?{,");
-		break;
-	case ARM_CC_LE:
-		// zf == 1 || nf != vf
-		close_type = 1;
-		rz_strbuf_setf(&op->esil, "zf,nf,vf,^,|,?{,");
-		break;
-	case ARM_CC_AL:
-		// always executed
-		break;
-	default:
-		break;
-	}
-	return close_cond[close_type];
-}
-
-/* arm64 */
-
-static const char *arg(RzAnalysis *a, csh *handle, cs_insn *insn, char *buf, int n) {
-	buf[0] = 0;
-	switch (insn->detail->arm.operands[n].type) {
-	case ARM_OP_REG:
-		if (ISSHIFTED(n)) {
-			sprintf(buf, "%u,%s,%s",
-				LSHIFT2(n),
-				rz_str_get_null(cs_reg_name(*handle,
-					insn->detail->arm.operands[n].reg)),
-				DECODE_SHIFT(n));
-		} else {
-			sprintf(buf, "%s",
-				rz_str_get_null(cs_reg_name(*handle,
-					insn->detail->arm.operands[n].reg)));
-		}
-		break;
-	case ARM_OP_IMM:
-		if (a->bits == 64) {
-			// 64bit only
-			sprintf(buf, "%" PFMT64d, (ut64)insn->detail->arm.operands[n].imm);
-		} else {
-			// 32bit only
-			sprintf(buf, "%" PFMT64d, (ut64)(ut32)insn->detail->arm.operands[n].imm);
-		}
-		break;
-	case ARM_OP_MEM:
-		break;
-	case ARM_OP_FP:
-		sprintf(buf, "%lf", insn->detail->arm.operands[n].fp);
-		break;
-	default:
-		break;
-	}
-	return buf;
-}
-
-#define ARG(x) arg(a, handle, insn, str[x], x)
-
-#define SHIFTED_REG64_APPEND(sb, n) shifted_reg64_append(sb, handle, insn, n)
-
-// do the sign extension here as well, but honestly this whole thing should maybe be redesigned
-static void shifted_reg64_append(RzStrBuf *sb, csh *handle, cs_insn *insn, int n) {
-	int signext = EXT64(n);
-	const char *rn;
-
-	if (HASMEMINDEX64(n)) {
-		rn = MEMINDEX64(n);
-	} else {
-		rn = REG64(n);
-	}
-
-	if (LSHIFT2_64(n)) {
-		if (insn->detail->arm64.operands[n].shift.type != ARM64_SFT_ASR) {
-			if (signext) {
-				rz_strbuf_appendf(sb, "%d,%d,%s,~,%s", LSHIFT2_64(n), signext, rn, DECODE_SHIFT64(n));
-			} else {
-				rz_strbuf_appendf(sb, "%d,%s,%s", LSHIFT2_64(n), rn, DECODE_SHIFT64(n));
-			}
-		} else {
-			/* ASR: add the missing ones if negative */
-			int index = LSHIFT2_64(n) - 1;
-			if (index < 0) {
-				return;
-			}
-			ut64 missing_ones = bitmask_by_width[index] << (REGSIZE64(n) * 8 - LSHIFT2_64(n));
-			if (signext) {
-				rz_strbuf_appendf(sb, "%d,%d,%s,~,%s,1,%d,%s,~,<<<,1,&,?{,%" PFMT64u ",}{,0,},|",
-					LSHIFT2_64(n), signext, rn, DECODE_SHIFT64(n), signext, REG64(n), (ut64)missing_ones);
-			} else {
-				rz_strbuf_appendf(sb, "%d,%s,%s,1,%s,<<<,1,&,?{,%" PFMT64u ",}{,0,},|",
-					LSHIFT2_64(n), rn, DECODE_SHIFT64(n), rn, (ut64)missing_ones);
-			}
-		}
-	} else if (signext) {
-		rz_strbuf_appendf(sb, "%d,%s,~", signext, rn);
-	} else {
-		rz_strbuf_appendf(sb, "%s", rn);
-	}
-}
-
-#define OPCALL(opchar)     arm64math(a, op, addr, buf, len, handle, insn, opchar, 0)
-#define OPCALL_NEG(opchar) arm64math(a, op, addr, buf, len, handle, insn, opchar, 1)
-
-// got rid of the opchar= pattern here because it caused missing operators to fail silently
-// and makes things more complicated with very little benefit
-static void arm64math(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, csh *handle, cs_insn *insn, const char *opchar, int negate) {
-	const char *r0 = REG64(0);
-	const char *r1 = REG64(1);
-
-	if (ISREG64(2)) {
-		if (LSHIFT2_64(2) || EXT64(2)) {
-			SHIFTED_REG64_APPEND(&op->esil, 2);
-			if (negate) {
-				rz_strbuf_appendf(&op->esil, ",-1,^");
-			}
-			rz_strbuf_appendf(&op->esil, ",%s,%s,%s,=", r1, opchar, r0);
-		} else {
-			const char *r2 = REG64(2);
-			if (negate) {
-				rz_strbuf_setf(&op->esil, "%s,-1,^,%s,%s,%s,=", r2, r1, opchar, r0);
-			} else {
-				rz_strbuf_setf(&op->esil, "%s,%s,%s,%s,=", r2, r1, opchar, r0);
-			}
-		}
-	} else {
-		ut64 i2 = IMM64(2) << LSHIFT2_64(2);
-		if (negate) {
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",-1,^,%s,%s,%s,=", i2, r1, opchar, r0);
-		} else {
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,%s,%s,=", i2, r1, opchar, r0);
-		}
-	}
-}
-
-static int analop64_esil(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, csh *handle, cs_insn *insn) {
-
-	const char *postfix = NULL;
-
-	rz_strbuf_init(&op->esil);
-	rz_strbuf_set(&op->esil, "");
-
-	postfix = arm_prefix_cond(op, insn->detail->arm64.cc);
-
-	switch (insn->id) {
-	case ARM64_INS_REV:
-		// these REV* instructions were almost right, except in the cases like rev x0, x0
-		// where the use of |= caused copies of the value to be erroneously present
-		{
-			const char *r0 = REG64(0);
-			const char *r1 = REG64(1);
-			int size = REGSIZE64(1);
-#if 0
-		rz_strbuf_setf (&op->esil,
-			"0,%s,=,"                        // dst = 0
-			"%d,"                            // initial counter = size
-			"DUP,"                           // counter: size -> 0 (repeat here)
-				"DUP,1,SWAP,-,8,*,"          // counter to bits in source
-					"DUP,0xff,<<,%s,&,>>,"   // src byte moved to LSB
-				"SWAP,%d,-,8,*,"             // invert counter, calc dst bit
-				"SWAP,<<,%s,|=,"             // shift left to there and insert
-			"4,REPEAT",                      // goto 5th instruction
-			r0, size, r1, size, r0);
-#endif
-			if (size == 8) {
-				rz_strbuf_setf(&op->esil,
-					"56,0xff,%s,&,<<,tmp,=,"
-					"48,0xff,8,%s,>>,&,<<,tmp,|=,"
-					"40,0xff,16,%s,>>,&,<<,tmp,|=,"
-					"32,0xff,24,%s,>>,&,<<,tmp,|=,"
-					"24,0xff,32,%s,>>,&,<<,tmp,|=,"
-					"16,0xff,40,%s,>>,&,<<,tmp,|=,"
-					"8,0xff,48,%s,>>,&,<<,tmp,|=,"
-					"0xff,56,%s,>>,&,tmp,|=,tmp,%s,=",
-					r1, r1, r1, r1,
-					r1, r1, r1, r1, r0);
-			} else {
-				rz_strbuf_setf(&op->esil,
-					"24,0xff,%s,&,<<,tmp,=,"
-					"16,0xff,8,%s,>>,&,<<,tmp,|=,"
-					"8,0xff,16,%s,>>,&,<<,tmp,|=,"
-					"0xff,24,%s,>>,&,tmp,|=,tmp,%s,=",
-					r1, r1, r1, r1, r0);
-			}
-			break;
-		}
-	case ARM64_INS_REV32: {
-		const char *r0 = REG64(0);
-		const char *r1 = REG64(1);
-		rz_strbuf_setf(&op->esil,
-			"24,0x000000ff000000ff,%s,&,<<,tmp,=,"
-			"16,0x000000ff000000ff,8,%s,>>,&,<<,tmp,|=,"
-			"8,0x000000ff000000ff,16,%s,>>,&,<<,tmp,|=,"
-			"0x000000ff000000ff,24,%s,>>,&,tmp,|=,tmp,%s,=",
-			r1, r1, r1, r1, r0);
-		break;
-	}
-	case ARM64_INS_REV16: {
-		const char *r0 = REG64(0);
-		const char *r1 = REG64(1);
-		rz_strbuf_setf(&op->esil,
-			"8,0xff00ff00ff00ff00,%s,&,>>,tmp,=,"
-			"8,0x00ff00ff00ff00ff,%s,&,<<,tmp,|=,tmp,%s,=",
-			r1, r1, r0);
-		break;
-	}
-	case ARM64_INS_ADR:
-		// TODO: must be 21bit signed
-		rz_strbuf_setf(&op->esil,
-			"%" PFMT64d ",%s,=", IMM64(1), REG64(0));
-		break;
-	case ARM64_INS_SMADDL: {
-		int size = REGSIZE64(1) * 8;
-		rz_strbuf_setf(&op->esil, "%d,%s,~,%d,%s,~,*,%s,+,%s,=",
-			size, REG64(2), size, REG64(1), REG64(3), REG64(0));
-		break;
-	}
-	case ARM64_INS_UMADDL:
-	case ARM64_INS_FMADD:
-	case ARM64_INS_MADD:
-		rz_strbuf_setf(&op->esil, "%s,%s,*,%s,+,%s,=",
-			REG64(2), REG64(1), REG64(3), REG64(0));
-		break;
-	case ARM64_INS_MSUB:
-		rz_strbuf_setf(&op->esil, "%s,%s,*,%s,-,%s,=",
-			REG64(2), REG64(1), REG64(3), REG64(0));
-		break;
-	case ARM64_INS_MNEG:
-		rz_strbuf_setf(&op->esil, "%s,%s,*,0,-,%s,=",
-			REG64(2), REG64(1), REG64(0));
-		break;
-	case ARM64_INS_ADD:
-	case ARM64_INS_ADC: // Add with carry.
-		// case ARM64_INS_ADCS: // Add with carry.
-		OPCALL("+");
-		break;
-	case ARM64_INS_SUB:
-		OPCALL("-");
-		break;
-	case ARM64_INS_SBC:
-		// TODO have to check this more, VEX does not work
-		rz_strbuf_setf(&op->esil, "%s,cf,+,%s,-,%s,=",
-			REG64(2), REG64(1), REG64(0));
-		break;
-	case ARM64_INS_SMULL: {
-		int size = REGSIZE64(1) * 8;
-		rz_strbuf_setf(&op->esil, "%d,%s,~,%d,%s,~,*,%s,=",
-			size, REG64(2), size, REG64(1), REG64(0));
-		break;
-	}
-	case ARM64_INS_MUL:
-		OPCALL("*");
-		break;
-	case ARM64_INS_AND:
-		OPCALL("&");
-		break;
-	case ARM64_INS_ORR:
-		OPCALL("|");
-		break;
-	case ARM64_INS_EOR:
-		OPCALL("^");
-		break;
-	case ARM64_INS_ORN:
-		OPCALL_NEG("|");
-		break;
-	case ARM64_INS_EON:
-		OPCALL_NEG("^");
-		break;
-	case ARM64_INS_LSR: {
-		const char *r0 = REG64(0);
-		const char *r1 = REG64(1);
-		const int size = REGSIZE64(0) * 8;
-
-		if (ISREG64(2)) {
-			if (LSHIFT2_64(2) || EXT64(2)) {
-				SHIFTED_REG64_APPEND(&op->esil, 2);
-				rz_strbuf_appendf(&op->esil, ",%d,%%,%s,>>,%s,=", size, r1, r0);
-			} else {
-				const char *r2 = REG64(2);
-				rz_strbuf_setf(&op->esil, "%d,%s,%%,%s,>>,%s,=", size, r2, r1, r0);
-			}
-		} else {
-			ut64 i2 = IMM64(2);
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,>>,%s,=", i2 % (ut64)size, r1, r0);
-		}
-		break;
-	}
-	case ARM64_INS_LSL: {
-		const char *r0 = REG64(0);
-		const char *r1 = REG64(1);
-		const int size = REGSIZE64(0) * 8;
-
-		if (ISREG64(2)) {
-			if (LSHIFT2_64(2) || EXT64(2)) {
-				SHIFTED_REG64_APPEND(&op->esil, 2);
-				rz_strbuf_appendf(&op->esil, ",%d,%%,%s,<<,%s,=", size, r1, r0);
-			} else {
-				const char *r2 = REG64(2);
-				rz_strbuf_setf(&op->esil, "%d,%s,%%,%s,<<,%s,=", size, r2, r1, r0);
-			}
-		} else {
-			ut64 i2 = IMM64(2);
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,<<,%s,=", i2 % (ut64)size, r1, r0);
-		}
-		break;
-	}
-	case ARM64_INS_ROR:
-		OPCALL(">>>");
-		break;
-	case ARM64_INS_NOP:
-		rz_strbuf_setf(&op->esil, ",");
-		break;
-	case ARM64_INS_FDIV:
-		break;
-	case ARM64_INS_SDIV: {
-		/* TODO: support WZR XZR to specify 32, 64bit op */
-		int size = REGSIZE64(1) * 8;
-		if (ISREG64(2)) {
-			rz_strbuf_setf(&op->esil, "%d,%s,~,%d,%s,~,~/,%s,=", size, REG64(2), size, REG64(1), REG64(0));
-		} else {
-			rz_strbuf_setf(&op->esil, "%d,%s,~,%d,%s,~,~/,%s,=", size, REG64(1), size, REG64(0), REG64(0));
-		}
-		break;
-	}
-	case ARM64_INS_UDIV:
-		/* TODO: support WZR XZR to specify 32, 64bit op */
-		if ISREG64 (2) {
-			rz_strbuf_setf(&op->esil, "%s,%s,/,%s,=", REG(2), REG64(1), REG64(0));
-		} else {
-			rz_strbuf_setf(&op->esil, "%s,%s,/=", REG64(1), REG64(0));
-		}
-		break;
-	case ARM64_INS_BR:
-		rz_strbuf_setf(&op->esil, "%s,pc,=", REG64(0));
-		break;
-	case ARM64_INS_B:
-		/* capstone precompute resulting address, using PC + IMM */
-		rz_strbuf_appendf(&op->esil, "%" PFMT64d ",pc,=", IMM64(0));
-		break;
-	case ARM64_INS_BL:
-		rz_strbuf_setf(&op->esil, "pc,lr,=,%" PFMT64d ",pc,=", IMM64(0));
-		break;
-	case ARM64_INS_BLR:
-		rz_strbuf_setf(&op->esil, "pc,lr,=,%s,pc,=", REG64(0));
-		break;
-	case ARM64_INS_CLZ:;
-		int size = 8 * REGSIZE64(0);
-
-		// expression is much more concise with GOTO, but GOTOs should be minimized
-		// rz_strbuf_setf (&op->esil, "%s,%s,=,0,DUP,%d,1,<<,%s,&,%d,>,&,?{,%s,=,}{,1,%s,<<=,1,+,4,GOTO,}",
-		//	REG64 (1), REG64 (0), size*8 - 1, REG64 (0), REG64 (0), REG64 (0));
-
-		/*
-		from https://en.wikipedia.org/wiki/Find_first_set modified for up to size 64
-		function clz3 (x)
-			if x = 0 return 32
-			n ← 0
-			if (x & 0xFFFF0000) = 0: n ← n + 16, x ← x << 16
-			if (x & 0xFF000000) = 0: n ← n +  8, x ← x <<  8
-			if (x & 0xF0000000) = 0: n ← n +  4, x ← x <<  4
-			if (x & 0xC0000000) = 0: n ← n +  2, x ← x <<  2
-			if (x & 0x80000000) = 0: n ← n +  1
-			return n
-		*/
-
-		const char *r0 = REG64(0);
-		const char *r1 = REG64(1);
-
-		if (size == 32) {
-			rz_strbuf_setf(&op->esil,
-				"%s,tmp,=,0,"
-				"tmp,0xffff0000,&,!,?{,16,tmp,<<=,16,+,},"
-				"tmp,0xff000000,&,!,?{,8,tmp,<<=,8,+,},"
-				"tmp,0xf0000000,&,!,?{,4,tmp,<<=,4,+,},"
-				"tmp,0xc0000000,&,!,?{,2,tmp,<<=,2,+,},"
-				"tmp,0x80000000,&,!,?{,1,+,},"
-				"%s,!,?{,32,%s,=,}{,%s,=,}",
-				r1, r1, r0, r0);
-		} else {
-			rz_strbuf_setf(&op->esil,
-				"%s,tmp,=,0,"
-				"tmp,0xffffffff00000000,&,!,?{,32,tmp,<<=,32,+,},"
-				"tmp,0xffff000000000000,&,!,?{,16,tmp,<<=,16,+,},"
-				"tmp,0xff00000000000000,&,!,?{,8,tmp,<<=,8,+,},"
-				"tmp,0xf000000000000000,&,!,?{,4,tmp,<<=,4,+,},"
-				"tmp,0xc000000000000000,&,!,?{,2,tmp,<<=,2,+,},"
-				"tmp,0x8000000000000000,&,!,?{,1,+,},"
-				"%s,!,?{,64,%s,=,}{,%s,=,}",
-				r1, r1, r0, r0);
-		}
-
-		break;
-	case ARM64_INS_LDRH:
-	case ARM64_INS_LDUR:
-	case ARM64_INS_LDURB:
-	case ARM64_INS_LDURH:
-	case ARM64_INS_LDR:
-	// case ARM64_INS_LDRSB:
-	// case ARM64_INS_LDRSH:
-	case ARM64_INS_LDRB:
-	// case ARM64_INS_LDRSW:
-	// case ARM64_INS_LDURSW:
-	case ARM64_INS_LDXR:
-	case ARM64_INS_LDXRB:
-	case ARM64_INS_LDXRH:
-	case ARM64_INS_LDAXR:
-	case ARM64_INS_LDAXRB:
-	case ARM64_INS_LDAXRH:
-	case ARM64_INS_LDAR:
-	case ARM64_INS_LDARB:
-	case ARM64_INS_LDARH: {
-		int size = REGSIZE64(0);
-		switch (insn->id) {
-		case ARM64_INS_LDRB:
-		case ARM64_INS_LDARB:
-		case ARM64_INS_LDAXRB:
-		case ARM64_INS_LDXRB:
-		case ARM64_INS_LDURB:
-			size = 1;
-			break;
-		case ARM64_INS_LDRH:
-		case ARM64_INS_LDARH:
-		case ARM64_INS_LDXRH:
-		case ARM64_INS_LDAXRH:
-		case ARM64_INS_LDURH:
-			size = 2;
-			break;
-		case ARM64_INS_LDRSW:
-		case ARM64_INS_LDURSW:
-			size = 4;
-			break;
-		default:
-			break;
-		}
-
-		if (ISMEM64(1)) {
-			if (HASMEMINDEX64(1)) {
-				if (LSHIFT2_64(1) || EXT64(1)) {
-					SHIFTED_REG64_APPEND(&op->esil, 1);
-					rz_strbuf_appendf(&op->esil, ",%s,+,[%d],%s,=", MEMBASE64(1), size, REG64(0));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%s,%s,+,[%d],%s,=",
-						MEMBASE64(1), MEMINDEX64(1), size, REG64(0));
-				}
-			} else {
-				// I really don't like the DUP / tmp approach but its better than doubling the calculation
-				if (LSHIFT2_64(1)) {
-					rz_strbuf_appendf(&op->esil, "%s,%d,%" PFMT64d ",%s,+",
-						MEMBASE64(1), LSHIFT2_64(1), MEMDISP64(1), DECODE_SHIFT64(1));
-				} else if ((int)MEMDISP64(1) < 0) {
-					rz_strbuf_appendf(&op->esil, "%" PFMT64d ",%s,-",
-						-(st64)MEMDISP64(1), MEMBASE64(1));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%" PFMT64d ",%s,+",
-						MEMDISP64(1), MEMBASE64(1));
-				}
-
-				rz_strbuf_append(&op->esil, ",DUP,tmp,=");
-
-				// I assume the DUPs here previously were to handle preindexing
-				// but it was never finished?
-				if (ISPREINDEX64()) {
-					rz_strbuf_appendf(&op->esil, ",tmp,%s,=", REG64(1));
-				}
-
-				rz_strbuf_appendf(&op->esil, ",[%d],%s,=", size, REG64(0));
-
-				if (ISPOSTINDEX64()) {
-					if (ISREG64(2)) { // not sure if register valued post indexing exists?
-						rz_strbuf_appendf(&op->esil, ",tmp,%s,+,%s,=", REG64(2), REG64(1));
-					} else {
-						rz_strbuf_appendf(&op->esil, ",tmp,%" PFMT64d ",+,%s,=", IMM64(2), REG64(1));
-					}
-				}
-			}
-			op->refptr = 4;
-		} else {
-			if (ISREG64(1)) {
-				if (OPCOUNT64() == 2) {
-					rz_strbuf_setf(&op->esil, "%s,[%d],%s,=",
-						REG64(1), size, REG64(0));
-				} else if (OPCOUNT64() == 3) {
-					/*
-						This seems like a capstone bug:
-						instructions like
-							ldr x16, [x13, x9]
-							ldrb w2, [x19, x23]
-						are not detected as ARM64_OP_MEM type and
-						fall in this case instead.
-					*/
-					if (ISREG64(2)) {
-						rz_strbuf_setf(&op->esil, "%s,%s,+,[%d],%s,=",
-							REG64(1), REG64(2), size, REG64(0));
-					}
-				}
-			} else {
-				rz_strbuf_setf(&op->esil, "%" PFMT64d ",[%d],%s,=",
-					IMM64(1), size, REG64(0));
-			}
-		}
-		break;
-	}
-	case ARM64_INS_LDRSB:
-	case ARM64_INS_LDRSH:
-	case ARM64_INS_LDRSW:
-	case ARM64_INS_LDURSB:
-	case ARM64_INS_LDURSH:
-	case ARM64_INS_LDURSW: {
-		// handle the sign extended instrs here
-		int size = REGSIZE64(0);
-		switch (insn->id) {
-		case ARM64_INS_LDRSB:
-		case ARM64_INS_LDURSB:
-			size = 1;
-			break;
-		case ARM64_INS_LDRSH:
-		case ARM64_INS_LDURSH:
-			size = 2;
-			break;
-		case ARM64_INS_LDRSW:
-		case ARM64_INS_LDURSW:
-			size = 4;
-			break;
-		default:
-			break;
-		}
-
-		if (ISMEM64(1)) {
-			if (HASMEMINDEX64(1)) {
-				if (LSHIFT2_64(1) || EXT64(1)) {
-					rz_strbuf_appendf(&op->esil, "%d,%s,", size * 8, MEMBASE64(1));
-					SHIFTED_REG64_APPEND(&op->esil, 1);
-					rz_strbuf_appendf(&op->esil, ",+,[%d],~,%s,=", size, REG64(0));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%d,%s,%s,+,[%d],~,%s,=",
-						size * 8, MEMBASE64(1), MEMINDEX64(1), size, REG64(0));
-				}
-			} else {
-				if (LSHIFT2_64(1)) {
-					rz_strbuf_appendf(&op->esil, "%d,%s,%d,%" PFMT64d ",%s",
-						size * 8, MEMBASE64(1), LSHIFT2_64(1), MEMDISP64(1), DECODE_SHIFT64(1));
-				} else if ((int)MEMDISP64(1) < 0) {
-					rz_strbuf_appendf(&op->esil, "%d,%" PFMT64d ",%s,-",
-						size * 8, -(st64)MEMDISP64(1), MEMBASE64(1));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%d,%" PFMT64d ",%s,+",
-						size * 8, MEMDISP64(1), MEMBASE64(1));
-				}
-
-				rz_strbuf_append(&op->esil, ",DUP,tmp,=");
-
-				// I assume the DUPs here previously were to handle preindexing
-				// but it was never finished?
-				if (ISPREINDEX64()) {
-					rz_strbuf_appendf(&op->esil, ",tmp,%s,=", REG64(1));
-				}
-
-				rz_strbuf_appendf(&op->esil, ",[%d],~,%s,=", size, REG64(0));
-
-				if (ISPOSTINDEX64()) {
-					if (ISREG64(2)) { // not sure if register valued post indexing exists?
-						rz_strbuf_appendf(&op->esil, ",tmp,%s,+,%s,=", REG64(2), REG64(1));
-					} else {
-						rz_strbuf_appendf(&op->esil, ",tmp,%" PFMT64d ",+,%s,=", IMM64(2), REG64(1));
-					}
-				}
-			}
-			op->refptr = 4;
-		} else {
-			if (ISREG64(1)) {
-				if (OPCOUNT64() == 2) {
-					rz_strbuf_setf(&op->esil, "%d,%s,[%d],~,%s,=",
-						size * 8, REG64(1), size, REG64(0));
-				} else if (OPCOUNT64() == 3) {
-					/*
-						This seems like a capstone bug:
-						instructions like
-							ldr x16, [x13, x9]
-							ldrb w2, [x19, x23]
-						are not detected as ARM64_OP_MEM type and
-						fall in this case instead.
-					*/
-					if (ISREG64(2)) {
-						rz_strbuf_setf(&op->esil, "%d,%s,%s,+,[%d],~,%s,=",
-							size * 8, REG64(1), REG64(2), size, REG64(0));
-					}
-				}
-			} else {
-				rz_strbuf_setf(&op->esil, "%d,%" PFMT64d ",[%d],~,%s,=",
-					size * 8, IMM64(1), size, REG64(0));
-			}
-		}
-		break;
-	}
-	case ARM64_INS_FCMP:
-	case ARM64_INS_CCMP:
-	case ARM64_INS_CCMN:
-	case ARM64_INS_TST: // cmp w8, 0xd
-	case ARM64_INS_CMP: // cmp w8, 0xd
-	case ARM64_INS_CMN: // cmp w8, 0xd
-	{
-		// update esil, cpu flags
-		int bits = arm64_reg_width(REGID64(0));
-		if (ISIMM64(1)) {
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,==,$z,zf,:=,%d,$s,nf,:=,%d,$b,!,cf,:=,%d,$o,vf,:=", IMM64(1) << LSHIFT2_64(1), REG64(0), bits - 1, bits, bits - 1);
-		} else {
-			// cmp w10, w11
-			SHIFTED_REG64_APPEND(&op->esil, 1);
-			rz_strbuf_appendf(&op->esil, ",%s,==,$z,zf,:=,%d,$s,nf,:=,%d,$b,!,cf,:=,%d,$o,vf,:=", REG64(0), bits - 1, bits, bits - 1);
-		}
-		break;
-	}
-	case ARM64_INS_FCSEL:
-	case ARM64_INS_CSEL: // csel Wd, Wn, Wm --> Wd := (cond) ? Wn : Wm
-		rz_strbuf_appendf(&op->esil, "%s,}{,%s,},%s,=", REG64(1), REG64(2), REG64(0));
-		postfix = "";
-		break;
-	case ARM64_INS_CSET: // cset Wd --> Wd := (cond) ? 1 : 0
-		rz_strbuf_appendf(&op->esil, "1,}{,0,},%s,=", REG64(0));
-		postfix = "";
-		break;
-	case ARM64_INS_CINC: // cinc Wd, Wn --> Wd := (cond) ? (Wn+1) : Wn
-		rz_strbuf_appendf(&op->esil, "1,%s,+,}{,%s,},%s,=", REG64(1), REG64(1), REG64(0));
-		postfix = "";
-		break;
-	case ARM64_INS_CSINC: // csinc Wd, Wn, Wm --> Wd := (cond) ? Wn : (Wm+1)
-		rz_strbuf_appendf(&op->esil, "%s,}{,1,%s,+,},%s,=", REG64(1), REG64(2), REG64(0));
-		postfix = "";
-		break;
-	case ARM64_INS_STXRB:
-	case ARM64_INS_STXRH:
-	case ARM64_INS_STXR: {
-		int size = REGSIZE64(1);
-		if (insn->id == ARM64_INS_STXRB) {
-			size = 1;
-		} else if (insn->id == ARM64_INS_STXRH) {
-			size = 2;
-		}
-		rz_strbuf_setf(&op->esil, "0,%s,=,%s,%s,%" PFMT64d ",+,=[%d]",
-			REG64(0), REG64(1), MEMBASE64(1), MEMDISP64(1), size);
-		break;
-	}
-	case ARM64_INS_STRB:
-	case ARM64_INS_STRH:
-	case ARM64_INS_STUR:
-	case ARM64_INS_STURB:
-	case ARM64_INS_STURH:
-	case ARM64_INS_STR: // str x6, [x6,0xf90]
-	{
-		int size = REGSIZE64(0);
-		if (insn->id == ARM64_INS_STRB || insn->id == ARM64_INS_STURB) {
-			size = 1;
-		} else if (insn->id == ARM64_INS_STRH || insn->id == ARM64_INS_STURH) {
-			size = 2;
-		}
-		if (ISMEM64(1)) {
-			if (HASMEMINDEX64(1)) {
-				if (LSHIFT2_64(1) || EXT64(1)) {
-					rz_strbuf_appendf(&op->esil, "%s,%s,", REG64(0), MEMBASE64(1));
-					SHIFTED_REG64_APPEND(&op->esil, 1);
-					rz_strbuf_appendf(&op->esil, ",+,=[%d]", size);
-				} else {
-					rz_strbuf_appendf(&op->esil, "%s,%s,%s,+,=[%d]",
-						REG64(0), MEMBASE64(1), MEMINDEX64(1), size);
-				}
-			} else {
-				if (LSHIFT2_64(1)) {
-					rz_strbuf_appendf(&op->esil, "%s,%s,%d,%" PFMT64d ",%s,+",
-						REG64(0), MEMBASE64(1), LSHIFT2_64(1), MEMDISP64(1), DECODE_SHIFT64(1));
-				} else if ((int)MEMDISP64(1) < 0) {
-					rz_strbuf_appendf(&op->esil, "%s,%" PFMT64d ",%s,-",
-						REG64(0), -(st64)MEMDISP64(1), MEMBASE64(1));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%s,%" PFMT64d ",%s,+",
-						REG64(0), MEMDISP64(1), MEMBASE64(1));
-				}
-
-				rz_strbuf_append(&op->esil, ",DUP,tmp,=");
-
-				// I assume the DUPs here previously were to handle preindexing
-				// but it was never finished?
-				if (ISPREINDEX64()) {
-					rz_strbuf_appendf(&op->esil, ",tmp,%s,=", REG64(1));
-				}
-
-				rz_strbuf_appendf(&op->esil, ",=[%d]", size);
-
-				if (ISPOSTINDEX64()) {
-					if (ISREG64(2)) { // not sure if register valued post indexing exists?
-						rz_strbuf_appendf(&op->esil, ",tmp,%s,+,%s,=", REG64(2), REG64(1));
-					} else {
-						rz_strbuf_appendf(&op->esil, ",tmp,%" PFMT64d ",+,%s,=", IMM64(2), REG64(1));
-					}
-				}
-			}
-			op->refptr = 4;
-		} else {
-			if (ISREG64(1)) {
-				if (OPCOUNT64() == 2) {
-					rz_strbuf_setf(&op->esil, "%s,%s,=[%d]",
-						REG64(0), REG64(1), size);
-				} else if (OPCOUNT64() == 3) {
-					/*
-						This seems like a capstone bug:
-						instructions like
-							ldr x16, [x13, x9]
-							ldrb w2, [x19, x23]
-						are not detected as ARM64_OP_MEM type and
-						fall in this case instead.
-					*/
-					if (ISREG64(2)) {
-						rz_strbuf_setf(&op->esil, "%s,%s,%s,+,=[%d]",
-							REG64(0), REG64(1), REG64(2), size);
-					}
-				}
-			} else {
-				rz_strbuf_setf(&op->esil, "%s,%" PFMT64d ",=[%d]",
-					REG64(0), IMM64(1), size);
-			}
-		}
-		break;
-	}
-	case ARM64_INS_BIC:
-		if (OPCOUNT64() == 2) {
-			if (REGSIZE64(0) == 4) {
-				rz_strbuf_appendf(&op->esil, "%s,0xffffffff,^,%s,&=", REG64(1), REG64(0));
-			} else {
-				rz_strbuf_appendf(&op->esil, "%s,0xffffffffffffffff,^,%s,&=", REG64(1), REG64(0));
-			}
-		} else {
-			if (REGSIZE64(0) == 4) {
-				rz_strbuf_appendf(&op->esil, "%s,0xffffffff,^,%s,&,%s,=", REG64(2), REG64(1), REG64(0));
-			} else {
-				rz_strbuf_appendf(&op->esil, "%s,0xffffffffffffffff,^,%s,&,%s,=", REG64(2), REG64(1), REG64(0));
-			}
-		}
-		break;
-	case ARM64_INS_CBZ:
-		rz_strbuf_setf(&op->esil, "%s,!,?{,%" PFMT64d ",pc,=,}",
-			REG64(0), IMM64(1));
-		break;
-	case ARM64_INS_CBNZ:
-		rz_strbuf_setf(&op->esil, "%s,?{,%" PFMT64d ",pc,=,}",
-			REG64(0), IMM64(1));
-		break;
-	case ARM64_INS_TBZ:
-		// tbnz x0, 4, label
-		// if ((1<<4) & x0) goto label;
-		rz_strbuf_setf(&op->esil, "%" PFMT64d ",1,<<,%s,&,!,?{,%" PFMT64d ",pc,=,}",
-			IMM64(1), REG64(0), IMM64(2));
-		break;
-	case ARM64_INS_TBNZ:
-		// tbnz x0, 4, label
-		// if ((1<<4) & x0) goto label;
-		rz_strbuf_setf(&op->esil, "%" PFMT64d ",1,<<,%s,&,?{,%" PFMT64d ",pc,=,}",
-			IMM64(1), REG64(0), IMM64(2));
-		break;
-	case ARM64_INS_STNP:
-	case ARM64_INS_STP: // stp x6, x7, [x6,0xf90]
-	{
-		int disp = (int)MEMDISP64(2);
-		char sign = disp >= 0 ? '+' : '-';
-		ut64 abs = disp >= 0 ? MEMDISP64(2) : -MEMDISP64(2);
-		int size = REGSIZE64(0);
-		// Pre-index case
-		if (ISPREINDEX64()) {
-			// "stp x2, x3, [x8, 0x20]!
-			// "32,x8,+=,x2,x8,=[8],x3,x8,8,+,=[8]",
-			rz_strbuf_setf(&op->esil,
-				"%" PFMT64d ",%s,%c=,%s,%s,=[%d],%s,%s,%d,+,=[%d]",
-				abs, MEMBASE64(2), sign,
-				REG64(0), MEMBASE64(2), size,
-				REG64(1), MEMBASE64(2), size, size);
-			// Post-index case
-		} else if (ISPOSTINDEX64()) {
-			int val = IMM64(3);
-			sign = val >= 0 ? '+' : '-';
-			abs = val >= 0 ? val : -val;
-			// "stp x4, x5, [x8], 0x10"
-			// "x4,x8,=[],x5,x8,8,+,=[],16,x8,+="
-			rz_strbuf_setf(&op->esil,
-				"%s,%s,=[%d],%s,%s,%d,+,=[%d],%" PFMT64d ",%s,%c=",
-				REG64(0), MEMBASE64(2), size,
-				REG64(1), MEMBASE64(2), size, size,
-				abs, MEMBASE64(2), sign);
-			// Everything else
-		} else {
-			rz_strbuf_setf(&op->esil,
-				"%s,%s,%" PFMT64d ",%c,=[%d],"
-				"%s,%s,%" PFMT64d ",%c,%d,+,=[%d]",
-				REG64(0), MEMBASE64(2), abs, sign, size,
-				REG64(1), MEMBASE64(2), abs, sign, size, size);
-		}
-	} break;
-	case ARM64_INS_LDP: // ldp x29, x30, [sp], 0x10
-	{
-		int disp = (int)MEMDISP64(2);
-		char sign = disp >= 0 ? '+' : '-';
-		ut64 abs = disp >= 0 ? MEMDISP64(2) : -MEMDISP64(2);
-		int size = REGSIZE64(0);
-		// Pre-index case
-		// x2,x8,32,+,=[8],x3,x8,32,+,8,+,=[8]
-		if (ISPREINDEX64()) {
-			// "ldp x0, x1, [x8, -0x10]!"
-			// 16,x8,-=,x8,[8],x0,=,x8,8,+,[8],x1,=
-			rz_strbuf_setf(&op->esil,
-				"%" PFMT64d ",%s,%c=,"
-				"%s,[%d],%s,=,"
-				"%d,%s,+,[%d],%s,=",
-				abs, MEMBASE64(2), sign,
-				MEMBASE64(2), size, REG64(0),
-				size, MEMBASE64(2), size, REG64(1));
-			// Post-index case
-		} else if (ISPOSTINDEX64()) {
-			int val = IMM64(3);
-			sign = val >= 0 ? '+' : '-';
-			abs = val >= 0 ? val : -val;
-			// ldp x4, x5, [x8], -0x10
-			// x8,[8],x4,=,x8,8,+,[8],x5,=,16,x8,+=
-			rz_strbuf_setf(&op->esil,
-				"%s,[%d],%s,=,"
-				"%s,%d,+,[%d],%s,=,"
-				"%" PFMT64d ",%s,%c=",
-				MEMBASE64(2), size, REG64(0),
-				MEMBASE64(2), size, size, REG64(1),
-				abs, MEMBASE64(2), sign);
-		} else {
-			rz_strbuf_setf(&op->esil,
-				"%" PFMT64d ",%s,%c,[%d],%s,=,"
-				"%d,%" PFMT64d ",%s,%c,+,[%d],%s,=",
-				abs, MEMBASE64(2), sign, size, REG64(0),
-				size, abs, MEMBASE64(2), sign, size, REG64(1));
-		}
-	} break;
-	case ARM64_INS_ADRP:
-		rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,=",
-			IMM64(1), REG64(0));
-		break;
-	case ARM64_INS_MOV:
-		if (ISREG64(1)) {
-			rz_strbuf_setf(&op->esil, "%s,%s,=", REG64(1), REG64(0));
-		} else {
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,=", IMM64(1), REG64(0));
-		}
-		break;
-	case ARM64_INS_EXTR:
-		// from VEX
-		/*
-			01 | t0 = GET:I64(x4)
-			02 | t1 = GET:I64(x0)
-			03 | t4 = Shr64(t1,0x20)
-			04 | t5 = Shl64(t0,0x20)
-			05 | t3 = Or64(t5,t4)
-			06 | PUT(x4) = t3
-		*/
-		rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,>>,%" PFMT64d ",%s,<<,|,%s,=",
-			IMM64(3), REG64(2), IMM64(3), REG64(1), REG64(0));
-		break;
-	case ARM64_INS_RBIT:
-		// this expression reverses the bits. it does. do not scroll right.
-		// Derived from VEX
-		rz_strbuf_setf(&op->esil, "0xffffffff00000000,0x20,0xffff0000ffff0000,0x10,0xff00ff00ff00ff00,0x8,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,<<,&,0x8,0xff00ff00ff00ff00,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,&,>>,|,<<,&,0x10,0xffff0000ffff0000,0xff00ff00ff00ff00,0x8,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,<<,&,0x8,0xff00ff00ff00ff00,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,&,>>,|,&,>>,|,<<,&,0x20,0xffffffff00000000,0xffff0000ffff0000,0x10,0xff00ff00ff00ff00,0x8,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,<<,&,0x8,0xff00ff00ff00ff00,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,&,>>,|,<<,&,0x10,0xffff0000ffff0000,0xff00ff00ff00ff00,0x8,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,<<,&,0x8,0xff00ff00ff00ff00,0xf0f0f0f0f0f0f0f0,0x4,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,<<,&,0x4,0xf0f0f0f0f0f0f0f0,0xcccccccccccccccc,0x2,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,<<,&,0x2,0xcccccccccccccccc,0xaaaaaaaaaaaaaaaa,0x1,%1$s,<<,&,0x1,0xaaaaaaaaaaaaaaaa,%1$s,&,>>,|,&,>>,|,&,>>,|,&,>>,|,&,>>,|,&,>>,|,%2$s,=",
-			REG64(1), REG64(0));
-		break;
-	case ARM64_INS_MVN:
-	case ARM64_INS_MOVN:
-		if (ISREG64(1)) {
-			rz_strbuf_setf(&op->esil, "%d,%s,-1,^,<<,%s,=", LSHIFT2_64(1), REG64(1), REG64(0));
-		} else {
-			rz_strbuf_setf(&op->esil, "%d,%" PFMT64d ",<<,-1,^,%s,=", LSHIFT2_64(1), IMM64(1), REG64(0));
-		}
-		break;
-	case ARM64_INS_MOVK: // movk w8, 0x1290
-	{
-		ut64 shift = LSHIFT2_64(1);
-		if (shift < 0) {
-			shift = 0;
-		} else if (shift > 48) {
-			shift = 48;
-		}
-		ut64 shifted_imm = IMM64(1) << shift;
-		ut64 mask = ~(0xffffLL << shift);
-
-		rz_strbuf_setf(&op->esil, "0x%" PFMT64x ",%s,&,%" PFMT64u ",|,%s,=",
-			mask,
-			REG64(0),
-			shifted_imm,
-			REG64(0));
-
-		break;
-	}
-	case ARM64_INS_MOVZ:
-		rz_strbuf_setf(&op->esil, "%" PFMT64u ",%s,=",
-			IMM64(1) << LSHIFT2_64(1),
-			REG64(0));
-		break;
-	/* ASR, SXTB, SXTH and SXTW are alias for SBFM */
-	case ARM64_INS_ASR: {
-		// OPCALL(">>>>");
-		const char *r0 = REG64(0);
-		const char *r1 = REG64(1);
-		const int size = REGSIZE64(0) * 8;
-
-		if (ISREG64(2)) {
-			if (LSHIFT2_64(2)) {
-				SHIFTED_REG64_APPEND(&op->esil, 2);
-				rz_strbuf_appendf(&op->esil, ",%d,%%,%s,>>>>,%s,=", size, r1, r0);
-			} else {
-				const char *r2 = REG64(2);
-				rz_strbuf_setf(&op->esil, "%d,%s,%%,%s,>>>>,%s,=", size, r2, r1, r0);
-			}
-		} else {
-			ut64 i2 = IMM64(2);
-			rz_strbuf_setf(&op->esil, "%" PFMT64d ",%s,>>>>,%s,=", i2 % (ut64)size, r1, r0);
-		}
-		break;
-	}
-	case ARM64_INS_SXTB:
-		if (arm64_reg_width(REGID64(0)) == 32) {
-			rz_strbuf_setf(&op->esil, "0xffffffff,8,0xff,%s,&,~,&,%s,=",
-				REG64(1), REG64(0));
-		} else {
-			rz_strbuf_setf(&op->esil, "8,0xff,%s,&,~,%s,=",
-				REG64(1), REG64(0));
-		}
-		break;
-	case ARM64_INS_SXTH: /* halfword */
-		if (arm64_reg_width(REGID64(0)) == 32) {
-			rz_strbuf_setf(&op->esil, "0xffffffff,16,0xffff,%s,&,~,&,%s,=",
-				REG64(1), REG64(0));
-		} else {
-			rz_strbuf_setf(&op->esil, "16,0xffff,%s,&,~,%s,=",
-				REG64(1), REG64(0));
-		}
-		break;
-	case ARM64_INS_SXTW: /* word */
-		rz_strbuf_setf(&op->esil, "32,0xffffffff,%s,&,~,%s,=",
-			REG64(1), REG64(0));
-		break;
-	case ARM64_INS_UXTB:
-		rz_strbuf_setf(&op->esil, "%s,0xff,&,%s,=", REG64(1), REG64(0));
-		break;
-	case ARM64_INS_UMULL:
-		rz_strbuf_setf(&op->esil, "%s,%s,*,%s,=", REG64(1), REG64(2), REG64(0));
-		break;
-	case ARM64_INS_UXTH:
-		rz_strbuf_setf(&op->esil, "%s,0xffff,&,%s,=", REG64(1), REG64(0));
-		break;
-	case ARM64_INS_RET:
-		rz_strbuf_setf(&op->esil, "lr,pc,=");
-		break;
-	case ARM64_INS_ERET:
-		rz_strbuf_setf(&op->esil, "lr,pc,=");
-		break;
-	case ARM64_INS_BFI: // bfi w8, w8, 2, 1
-	case ARM64_INS_BFXIL: {
-		if (OPCOUNT64() >= 3 && ISIMM64(3) && IMM64(3) > 0) {
-			ut64 mask = bitmask_by_width[IMM64(3) - 1];
-			ut64 shift = IMM64(2);
-			ut64 notmask = ~(mask << shift);
-			// notmask,dst,&,lsb,mask,src,&,<<,|,dst,=
-			rz_strbuf_setf(&op->esil, "%" PFMT64u ",%s,&,%" PFMT64u ",%" PFMT64u ",%s,&,<<,|,%s,=",
-				notmask, REG64(0), shift, mask, REG64(1), REG64(0));
-		}
-		break;
-	}
-	case ARM64_INS_SBFIZ:
-		if (IMM64(3) > 0 && IMM64(3) <= 64 - IMM64(2)) {
-			rz_strbuf_appendf(&op->esil, "%" PFMT64d ",%" PFMT64d ",%s,%" PFMT64u ",&,~,<<,%s,=",
-				IMM64(2), IMM64(3), REG64(1), (ut64)bitmask_by_width[IMM64(3) - 1], REG64(0));
-		}
-		break;
-	case ARM64_INS_UBFIZ:
-		if (IMM64(3) > 0 && IMM64(3) <= 64 - IMM64(2)) {
-			rz_strbuf_appendf(&op->esil, "%" PFMT64d ",%s,%" PFMT64u ",&,<<,%s,=",
-				IMM64(2), REG64(1), (ut64)bitmask_by_width[IMM64(3) - 1], REG64(0));
-		}
-		break;
-	case ARM64_INS_SBFX:
-		if (IMM64(3) > 0 && IMM64(3) <= 64 - IMM64(2)) {
-			rz_strbuf_appendf(&op->esil, "%" PFMT64d ",%" PFMT64d ",%s,%" PFMT64d ",%" PFMT64u ",<<,&,>>,~,%s,=",
-				IMM64(3), IMM64(2), REG64(1), IMM64(2), (ut64)bitmask_by_width[IMM64(3) - 1], REG64(0));
-		}
-		break;
-	case ARM64_INS_UBFX:
-		if (IMM64(3) > 0 && IMM64(3) <= 64 - IMM64(2)) {
-			rz_strbuf_appendf(&op->esil, "%" PFMT64d ",%s,%" PFMT64d ",%" PFMT64u ",<<,&,>>,%s,=",
-				IMM64(2), REG64(1), IMM64(2), (ut64)bitmask_by_width[IMM64(3) - 1], REG64(0));
-		}
-		break;
-	case ARM64_INS_NEG:
-#if CS_API_MAJOR > 3
-	case ARM64_INS_NEGS:
-#endif
-		if (LSHIFT2_64(1)) {
-			SHIFTED_REG64_APPEND(&op->esil, 1);
-		} else {
-			rz_strbuf_appendf(&op->esil, "%s", REG64(1));
-		}
-		rz_strbuf_appendf(&op->esil, ",0,-,%s,=", REG64(0));
-		break;
-	case ARM64_INS_SVC:
-		rz_strbuf_setf(&op->esil, "%" PFMT64u ",$", IMM64(0));
-		break;
-	}
-
-	rz_strbuf_append(&op->esil, postfix);
-
-	return 0;
-}
-
-#define MATH32(opchar)     arm32math(a, op, addr, buf, len, handle, insn, pcdelta, str, opchar, 0)
-#define MATH32_NEG(opchar) arm32math(a, op, addr, buf, len, handle, insn, pcdelta, str, opchar, 1)
-#define MATH32AS(opchar)   arm32mathaddsub(a, op, addr, buf, len, handle, insn, pcdelta, str, opchar)
-
-static void arm32math(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, csh *handle, cs_insn *insn, int pcdelta, char (*str)[32], const char *opchar, int negate) {
-	const char *dest = ARG(0);
-	const char *op1;
-	const char *op2;
-	bool rotate_imm = OPCOUNT() > 3;
-	if (OPCOUNT() > 2) {
-		op1 = ARG(1);
-		op2 = ARG(2);
-	} else {
-		op1 = dest;
-		op2 = ARG(1);
-	}
-	// right operand
-	if (rotate_imm) {
-		rz_strbuf_appendf(&op->esil, "%s,", ARG(3));
-	}
-	if (!strcmp(op2, "pc")) {
-		rz_strbuf_appendf(&op->esil, "%d,$$,+", pcdelta);
-	} else {
-		rz_strbuf_appendf(&op->esil, "%s", op2);
-	}
-	if (rotate_imm) {
-		rz_strbuf_appendf(&op->esil, ",>>>");
-	}
-	if (negate) {
-		rz_strbuf_appendf(&op->esil, ",-1,^");
-	}
-	if (!strcmp(op1, "pc")) {
-		rz_strbuf_appendf(&op->esil, ",%d,$$,+,%s,0xffffffff,&,%s,=", pcdelta, opchar, dest);
-	} else {
-		if (ISSHIFTED(1)) {
-			rz_strbuf_appendf(&op->esil, ",0xffffffff,&,%s,=", dest);
-		} else {
-			rz_strbuf_appendf(&op->esil, ",%s,%s,0xffffffff,&,%s,=", op1, opchar, dest);
-		}
-	}
-}
-
-static void arm32mathaddsub(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, csh *handle, cs_insn *insn, int pcdelta, char (*str)[32], const char *opchar) {
-	const char *dst = ARG(0);
-	const char *src;
-	bool noflags = false;
-	if (!strcmp(dst, "pc")) { // this is because strbuf_prepend doesn't exist and E_TOO_LAZY
-		//		rz_strbuf_append (&op->esil, "$$,pc,=,");
-		noflags = true;
-	}
-	if (OPCOUNT() == 3) {
-		rz_strbuf_appendf(&op->esil, "%s,0xffffffff,&,%s,=,", ARG(1), dst);
-		src = ARG(2);
-	} else {
-		//		src = (!strcmp (ARG(1), "pc"))? "$$": ARG(1);
-		src = ARG(1);
-	}
-	rz_strbuf_appendf(&op->esil, "%s,%s,%s,0xffffffff,&,%s,=", src, dst, opchar, dst);
-	if (noflags) {
-		return;
-	}
-	rz_strbuf_appendf(&op->esil, ",$z,zf,:=,%s,cf,:=,vf,=,0,nf,=",
-		(!strcmp(opchar, "+") ? "30,$c,31,$c,^,31,$c" : "30,$c,31,$c,^,32,$b"));
-}
-
-static int analop_esil(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, csh *handle, cs_insn *insn, bool thumb) {
-	int i;
-	const char *postfix = NULL;
-	char str[32][32];
-	int msr_flags;
-	int pcdelta = (thumb ? 4 : 8);
-	ut32 mask = UT32_MAX;
-	int str_ldr_bytes = 4;
-	unsigned int width = 0;
-
-	rz_strbuf_init(&op->esil);
-	rz_strbuf_set(&op->esil, "");
-	postfix = arm_prefix_cond(op, insn->detail->arm.cc);
-
-	switch (insn->id) {
-	case ARM_INS_CLZ:
-		rz_strbuf_appendf(&op->esil, "%s,!,?{,32,%s,=,BREAK,},0,%s,=,%s,%s,<<,0x80000000,&,!,?{,1,%s,+=,11,GOTO,}", REG(1), REG(0), REG(0), REG(0), REG(1), REG(0));
-		break;
-	case ARM_INS_IT:
-		rz_strbuf_appendf(&op->esil, "2,$$,+,pc,=");
-		break;
-	case ARM_INS_BKPT:
-		rz_strbuf_setf(&op->esil, "%d,%d,TRAP", IMM(0), IMM(0));
-		break;
-	case ARM_INS_NOP:
-		rz_strbuf_setf(&op->esil, ",");
-		break;
-	case ARM_INS_BL:
-	case ARM_INS_BLX:
-		rz_strbuf_appendf(&op->esil, "pc,%d,+,lr,=,", thumb);
-		/* fallthrough */
-	case ARM_INS_BX:
-	case ARM_INS_BXJ:
-	case ARM_INS_B:
-		if (ISREG(0) && REGID(0) == ARM_REG_PC) {
-			rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",pc,=", (addr & ~3LL) + pcdelta);
-		} else {
-			if (ISIMM(0)) {
-				rz_strbuf_appendf(&op->esil, "%s,pc,=", ARG(0));
-			} else {
-				rz_strbuf_appendf(&op->esil, "%d,%s,-,pc,=", thumb, ARG(0));
-			}
-		}
-		break;
-	case ARM_INS_UDF:
-		rz_strbuf_setf(&op->esil, "%s,TRAP", ARG(0));
-		break;
-	case ARM_INS_SADD16:
-	case ARM_INS_SADD8:
-		MATH32AS("+");
-		break;
-	case ARM_INS_ADDW:
-	case ARM_INS_ADD:
-		MATH32("+");
-		break;
-	case ARM_INS_ADC:
-		if (OPCOUNT() == 2) {
-			rz_strbuf_appendf(&op->esil, "cf,%s,+=,%s,%s,+=", ARG(0), ARG(1), ARG(0));
-		} else {
-			rz_strbuf_appendf(&op->esil, "cf,%s,+=,%s,%s,+,%s,+=", ARG(0), ARG(2), ARG(1), ARG(0));
-		}
-		break;
-	case ARM_INS_SSUB16:
-	case ARM_INS_SSUB8:
-		MATH32AS("-");
-		break;
-	case ARM_INS_SUBW:
-	case ARM_INS_SUB:
-		MATH32("-");
-		break;
-	case ARM_INS_SBC:
-		if (OPCOUNT() == 2) {
-			rz_strbuf_appendf(&op->esil, "cf,%s,-=,%s,%s,-=", ARG(0), ARG(1), ARG(0));
-		} else {
-			rz_strbuf_appendf(&op->esil, "cf,%s,-=,%s,%s,+,%s,-=", ARG(0), ARG(2), ARG(1), ARG(0));
-		}
-		break;
-	case ARM_INS_MUL:
-		MATH32("*");
-		break;
-	case ARM_INS_AND:
-		MATH32("&");
-		break;
-	case ARM_INS_ORR:
-		MATH32("|");
-		break;
-	case ARM_INS_EOR:
-		MATH32("^");
-		break;
-	case ARM_INS_ORN:
-		MATH32_NEG("|");
-		break;
-	case ARM_INS_LSR:
-		if (insn->detail->arm.update_flags) {
-			if (OPCOUNT() == 2) {
-				rz_strbuf_appendf(&op->esil, "%s,!,!,?{,%s,1,%s,-,0x1,<<,&,!,!,cf,:=,},", ARG(1), ARG(0), ARG(1));
-			} else {
-				rz_strbuf_appendf(&op->esil, "%s,!,!,?{,%s,1,%s,-,0x1,<<,&,!,!,cf,:=,},", ARG(2), ARG(1), ARG(2));
-			}
-		}
-		MATH32(">>");
-		break;
-	case ARM_INS_LSL:
-		if (insn->detail->arm.update_flags) {
-			if (OPCOUNT() == 2) {
-				rz_strbuf_appendf(&op->esil, "%s,!,!,?{,%s,32,-,%s,>>,cf,:=,},", ARG(1), ARG(1), ARG(0));
-			} else {
-				rz_strbuf_appendf(&op->esil, "%s,!,!,?{,%s,32,-,%s,>>,cf,:=,},", ARG(2), ARG(2), ARG(1));
-			}
-		}
-		MATH32("<<");
-		break;
-	case ARM_INS_SVC:
-		rz_strbuf_setf(&op->esil, "%s,$", ARG(0));
-		break;
-	case ARM_INS_PUSH:
-#if 0
-PUSH { r4, r5, r6, r7, lr }
-4,sp,-=,lr,sp,=[4],
-4,sp,-=,r7,sp,=[4],
-4,sp,-=,r6,sp,=[4],
-4,sp,-=,r5,sp,=[4],
-4,sp,-=,r4,sp,=[4]
-
-20,sp,-=,lr,r7,r6,r5,r4,5,sp,=[*]
-#endif
-		rz_strbuf_appendf(&op->esil, "%d,sp,-=,",
-			4 * insn->detail->arm.op_count);
-		for (i = insn->detail->arm.op_count; i > 0; i--) {
-			rz_strbuf_appendf(&op->esil, "%s,", REG(i - 1));
-		}
-		rz_strbuf_appendf(&op->esil, "%d,sp,=[*]",
-			insn->detail->arm.op_count);
-		break;
-	case ARM_INS_STMDA:
-	case ARM_INS_STMDB:
-	case ARM_INS_STM:
-	case ARM_INS_STMIB: {
-		int direction = (insn->id == ARM_INS_STMDA || insn->id == ARM_INS_STMDB ? -1 : 1);
-		int offset = direction > 0 ? -1 : -insn->detail->arm.op_count;
-		if (insn->id == ARM_INS_STMDA || insn->id == ARM_INS_STMIB) {
-			offset++;
-		}
-		for (i = 1; i < insn->detail->arm.op_count; i++) {
-			rz_strbuf_appendf(&op->esil, "%s,%s,%d,+,=[4],",
-				REG(i), ARG(0), (i + offset) * 4);
-		}
-		if (insn->detail->arm.writeback == true) { // writeback, reg should be incremented
-			rz_strbuf_appendf(&op->esil, "%d,%s,+=,",
-				direction * (insn->detail->arm.op_count - 1) * 4, ARG(0));
-		}
-		break;
-	}
-	case ARM_INS_VSTMIA:
-		rz_strbuf_set(&op->esil, "");
-		width = 0;
-		for (i = 1; i < insn->detail->arm.op_count; i++) {
-			rz_strbuf_appendf(&op->esil, "%s,%d,%s,+,=[%d],",
-				REG(i), width, ARG(0), REGSIZE32(i));
-			width += REGSIZE32(i);
-		}
-		// increment if writeback
-		if (insn->detail->arm.writeback) {
-			rz_strbuf_appendf(&op->esil, "%d,%s,+=,", width, ARG(0));
-		}
-		break;
-	case ARM_INS_VSTMDB:
-		rz_strbuf_set(&op->esil, "");
-		width = 0;
-		for (i = insn->detail->arm.op_count - 1; i > 0; i--) {
-			width += REGSIZE32(i);
-			rz_strbuf_appendf(&op->esil, "%s,%d,%s,-,=[%d],",
-				REG(i), width, ARG(0), REGSIZE32(i));
-		}
-		// decrement writeback is mandatory for VSTMDB
-		rz_strbuf_appendf(&op->esil, "%d,%s,-=,", width, ARG(0));
-		break;
-	case ARM_INS_VLDMIA:
-		rz_strbuf_set(&op->esil, "");
-		width = 0;
-		for (i = 1; i < insn->detail->arm.op_count; i++) {
-			rz_strbuf_appendf(&op->esil, "%d,%s,+,[%d],%s,=,",
-				width, ARG(0), REGSIZE32(i), REG(i));
-			width += REGSIZE32(i);
-		}
-		// increment if writeback
-		if (insn->detail->arm.writeback) {
-			rz_strbuf_appendf(&op->esil, "%d,%s,+=,", width, ARG(0));
-		}
-		break;
-	case ARM_INS_VLDMDB:
-		rz_strbuf_set(&op->esil, "");
-		width = 0;
-		for (i = insn->detail->arm.op_count - 1; i > 0; i--) {
-			width += REGSIZE32(i);
-			rz_strbuf_appendf(&op->esil, "%d,%s,-,[%d],%s,=,",
-				width, ARG(0), REGSIZE32(i), REG(i));
-		}
-		// decrement writeback is mandatory for VLDMDB
-		rz_strbuf_appendf(&op->esil, "%d,%s,-=,", width, ARG(0));
-		break;
-	case ARM_INS_ASR:
-		// suffix 'S' forces conditional flag to be updated
-		if (insn->detail->arm.update_flags) {
-			if (OPCOUNT() == 2) {
-				rz_strbuf_appendf(&op->esil, "%s,!,!,?{,%s,1,%s,-,0x1,<<,&,!,!,cf,:=,},", ARG(1), ARG(0), ARG(1));
-			} else if (OPCOUNT() == 3) {
-				rz_strbuf_appendf(&op->esil, "%s,!,!,?{,%s,1,%s,-,0x1,<<,&,!,!,cf,:=,},", ARG(2), ARG(1), ARG(2));
-			}
-		}
-		if (OPCOUNT() == 2) {
-			if (ISSHIFTED(1)) {
-				rz_strbuf_appendf(&op->esil, "%s,%s,=", ARG(1), ARG(0));
-			} else {
-				rz_strbuf_appendf(&op->esil, "%s,%s,>>>>,%s,=", ARG(1), ARG(0), ARG(0));
-			}
-		} else if (OPCOUNT() == 3) {
-			rz_strbuf_appendf(&op->esil, "%s,%s,>>>>,%s,=", ARG(2), ARG(1), ARG(0));
-		}
-		break;
-	case ARM_INS_POP:
-#if 0
-POP { r4,r5, r6}
-r6,r5,r4,3,sp,[*],12,sp,+=
-#endif
-		for (i = insn->detail->arm.op_count; i > 0; i--) {
-			rz_strbuf_appendf(&op->esil, "%s,", REG(i - 1));
-		}
-		rz_strbuf_appendf(&op->esil, "%d,sp,[*],",
-			insn->detail->arm.op_count);
-		rz_strbuf_appendf(&op->esil, "%d,sp,+=",
-			4 * insn->detail->arm.op_count);
-		break;
-	case ARM_INS_LDMDA:
-	case ARM_INS_LDMDB:
-	case ARM_INS_LDM:
-	case ARM_INS_LDMIB: {
-		int direction = (insn->id == ARM_INS_LDMDA || insn->id == ARM_INS_LDMDB) ? -1 : 1;
-		int offset = direction > 0 ? -1 : -insn->detail->arm.op_count;
-		if (insn->id == ARM_INS_LDMDA || insn->id == ARM_INS_LDMIB) {
-			offset++;
-		}
-		for (i = 1; i < insn->detail->arm.op_count; i++) {
-			rz_strbuf_appendf(&op->esil, "%s,%d,+,[4],%s,=,", ARG(0), (i + offset) * 4, REG(i));
-		}
-		if (insn->detail->arm.writeback) {
-			rz_strbuf_appendf(&op->esil, "%d,%s,+=,",
-				direction * (insn->detail->arm.op_count - 1) * 4, ARG(0));
-		}
-		break;
-	}
-	case ARM_INS_CMP:
-		rz_strbuf_appendf(&op->esil, "%s,%s,==", ARG(1), ARG(0));
-		break;
-	case ARM_INS_CMN:
-		rz_strbuf_appendf(&op->esil, "%s,%s,^,!,!,zf,=", ARG(1), ARG(0));
-		break;
-	case ARM_INS_MOVT:
-		rz_strbuf_appendf(&op->esil, "16,%s,<<,%s,|=", ARG(1), REG(0));
-		break;
-	case ARM_INS_ADR:
-		rz_strbuf_appendf(&op->esil, "%d,$$,+,%s,+,0xfffffffc,&,%s,=",
-			pcdelta, ARG(1), REG(0));
-		break;
-	case ARM_INS_MOV:
-	case ARM_INS_VMOV:
-	case ARM_INS_MOVW:
-		rz_strbuf_appendf(&op->esil, "%s,%s,=", ARG(1), REG(0));
-		break;
-	case ARM_INS_CBZ:
-		rz_strbuf_appendf(&op->esil, "%s,!,?{,%" PFMT32u ",pc,=,}",
-			REG(0), IMM(1));
-		break;
-	case ARM_INS_CBNZ:
-		rz_strbuf_appendf(&op->esil, "%s,?{,%" PFMT32u ",pc,=,}",
-			REG(0), IMM(1));
-		break;
-		// Encapsulated STR/H/B into a code section
-	case ARM_INS_STRT:
-	case ARM_INS_STR:
-	case ARM_INS_STRHT:
-	case ARM_INS_STRH:
-	case ARM_INS_STRBT:
-	case ARM_INS_STRB:
-	case ARM_INS_STRD:
-		// case ARM_INS_STLXRB: // capstone has no STLXR?
-		switch (insn->id) {
-		case ARM_INS_STRD:
-			str_ldr_bytes = 8; // just an indication, won't be used in esil code
-			break;
-		case ARM_INS_STRHT:
-		case ARM_INS_STRH:
-			str_ldr_bytes = 2;
-			break;
-		case ARM_INS_STRBT:
-		case ARM_INS_STRB:
-			str_ldr_bytes = 1;
-			break;
-		default:
-			str_ldr_bytes = 4;
-		}
-		if (OPCOUNT() == 2) {
-			if (ISMEM(1) && !HASMEMINDEX(1)) {
-				int disp = MEMDISP(1);
-				char sign = disp >= 0 ? '+' : '-';
-				disp = disp >= 0 ? disp : -disp;
-				rz_strbuf_appendf(&op->esil, "%s,0x%x,%s,%c,0xffffffff,&,=[%d]",
-					REG(0), disp, MEMBASE(1), sign, str_ldr_bytes);
-				if (insn->detail->arm.writeback) {
-					rz_strbuf_appendf(&op->esil, ",%d,%s,%c,%s,=",
-						disp, MEMBASE(1), sign, MEMBASE(1));
-				}
-			}
-			if (HASMEMINDEX(1)) { // e.g. 'str r2, [r3, r1]'
-				if (ISSHIFTED(1)) { // e.g. 'str r2, [r3, r1, lsl 4]'
-					switch (SHIFTTYPE(1)) {
-					case ARM_SFT_LSL:
-						rz_strbuf_appendf(&op->esil, "%s,%s,%d,%s,<<,+,0xffffffff,&,=[%d]",
-							REG(0), MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), str_ldr_bytes);
-						if (insn->detail->arm.writeback) { // e.g. 'str r2, [r3, r1, lsl 4]!'
-							rz_strbuf_appendf(&op->esil, ",%s,%d,%s,<<,+,%s,=",
-								MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), MEMBASE(1));
-						}
-						break;
-					case ARM_SFT_LSR:
-						rz_strbuf_appendf(&op->esil, "%s,%s,%d,%s,>>,+,0xffffffff,&,=[%d]",
-							REG(0), MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), str_ldr_bytes);
-						if (insn->detail->arm.writeback) {
-							rz_strbuf_appendf(&op->esil, ",%s,%d,%s,>>,+,%s,=",
-								MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), MEMBASE(1));
-						}
-						break;
-					case ARM_SFT_ASR:
-						rz_strbuf_appendf(&op->esil, "%s,%s,%d,%s,>>>>,+,0xffffffff,&,=[%d]",
-							REG(0), MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), str_ldr_bytes);
-						if (insn->detail->arm.writeback) {
-							rz_strbuf_appendf(&op->esil, ",%s,%d,%s,>>>>,+,%s,=",
-								MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), MEMBASE(1));
-						}
-						break;
-					case ARM_SFT_ROR:
-						rz_strbuf_appendf(&op->esil, "%s,%s,%d,%s,>>>,+,0xffffffff,&,=[%d]",
-							REG(0), MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), str_ldr_bytes);
-						if (insn->detail->arm.writeback) {
-							rz_strbuf_appendf(&op->esil, ",%s,%d,%s,>>>,+,%s,=",
-								MEMBASE(1), SHIFTVALUE(1), MEMINDEX(1), MEMBASE(1));
-						}
-						break;
-					case ARM_SFT_RRX: // ROR with single bit shift, using previous cf rather than new cf
-						// TODO: r2 doesn't mark this as a shift, it falls through to no shift
-						break;
-					default:
-						// Hopefully nothing here
-						break;
-					}
-				} else { // No shift
-					rz_strbuf_appendf(&op->esil, "%s,%s,%s,+,0xffffffff,&,=[%d]",
-						REG(0), MEMINDEX(1), MEMBASE(1), str_ldr_bytes);
-					if (insn->detail->arm.writeback) {
-						rz_strbuf_appendf(&op->esil, ",%s,%s,+,%s,=",
-							MEMINDEX(1), MEMBASE(1), MEMBASE(1));
-					}
-				}
-			}
-		}
-		if (OPCOUNT() == 3) { // e.g. 'str r2, [r3], 4
-			if (ISIMM(2)) { // e.g. 'str r2, [r3], 4
-				rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%d,%s,+=",
-					REG(0), MEMBASE(1), str_ldr_bytes, IMM(2), MEMBASE(1));
-			}
-			if (ISREG(2)) { // e.g. 'str r2, [r3], r1
-				if (ISSHIFTED(2)) { // e.g. 'str r2, [r3], r1, lsl 4'
-					switch (SHIFTTYPE(2)) {
-					case ARM_SFT_LSL:
-						rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,%d,%s,<<,+,%s,=",
-							REG(0), MEMBASE(1), str_ldr_bytes, MEMBASE(1), SHIFTVALUE(2), REG(2), MEMBASE(1));
-						break;
-					case ARM_SFT_LSR:
-						rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,%d,%s,>>,+,%s,=",
-							REG(0), MEMBASE(1), str_ldr_bytes, MEMBASE(1), SHIFTVALUE(2), REG(2), MEMBASE(1));
-						break;
-					case ARM_SFT_ASR:
-						rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,%d,%s,>>>>,+,%s,=",
-							REG(0), MEMBASE(1), str_ldr_bytes, MEMBASE(1), SHIFTVALUE(2), REG(2), MEMBASE(1));
-						break;
-					case ARM_SFT_ROR:
-						rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,%d,%s,>>>,+,%s,=",
-							REG(0), MEMBASE(1), str_ldr_bytes, MEMBASE(1), SHIFTVALUE(2), REG(2), MEMBASE(1));
-						break;
-					case ARM_SFT_RRX:
-						// TODO
-						break;
-					default:
-						// Hopefully nothing here
-						break;
-					}
-				} else { // No shift
-					rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,%s,+=",
-						REG(0), MEMBASE(1), str_ldr_bytes, REG(2), MEMBASE(1));
-				}
-			}
-			if (ISREG(1) && str_ldr_bytes == 8) { // e.g. 'strd r2, r3, [r4]', normally should be the only case for ISREG(1).
-				if (!HASMEMINDEX(2)) {
-					int disp = MEMDISP(2);
-					char sign = disp >= 0 ? '+' : '-';
-					disp = disp >= 0 ? disp : -disp;
-					rz_strbuf_appendf(&op->esil, "%s,%d,%s,%c,0xffffffff,&,=[4],%s,4,%d,+,%s,%c,0xffffffff,&,=[4]",
-						REG(0), disp, MEMBASE(2), sign, REG(1), disp, MEMBASE(2), sign);
-					if (insn->detail->arm.writeback) {
-						rz_strbuf_appendf(&op->esil, ",%d,%s,%c,%s,=",
-							disp, MEMBASE(2), sign, MEMBASE(2));
-					}
-				} else {
-					if (ISSHIFTED(2)) {
-						// it seems strd does not support SHIFT which is good, but have a check nonetheless
-					} else {
-						rz_strbuf_appendf(&op->esil, "%s,%s,%s,+,0xffffffff,&,=[4],%s,4,%s,+,%s,+,0xffffffff,&,=[4]",
-							REG(0), MEMINDEX(2), MEMBASE(2), REG(1), MEMINDEX(2), MEMBASE(2));
-						if (insn->detail->arm.writeback) {
-							rz_strbuf_appendf(&op->esil, ",%s,%s,+,%s,=",
-								MEMINDEX(2), MEMBASE(2), MEMBASE(2));
-						}
-					}
-				}
-			}
-		}
-		if (OPCOUNT() == 4) { // e.g. 'strd r2, r3, [r4], 4' or 'strd r2, r3, [r4], r5'
-			if (ISIMM(3)) { // e.g. 'strd r2, r3, [r4], 4'
-				rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,4,%s,+,0xffffffff,&,=[%d],%d,%s,+=,",
-					REG(0), MEMBASE(2), str_ldr_bytes, REG(1), MEMBASE(2), str_ldr_bytes, IMM(3), MEMBASE(2));
-			}
-			if (ISREG(3)) { // e.g. 'strd r2, r3, [r4], r5'
-				if (ISSHIFTED(3)) {
-					// same as above
-				} else {
-					rz_strbuf_appendf(&op->esil, "%s,%s,0xffffffff,&,=[%d],%s,4,%s,+,0xffffffff,&,=[%d],%s,%s,+=",
-						REG(0), MEMBASE(2), str_ldr_bytes, REG(1), MEMBASE(2), str_ldr_bytes, REG(3), MEMBASE(2));
-				}
-			}
-		}
-		break;
-	case ARM_INS_TST:
-		rz_strbuf_appendf(&op->esil, "0,%s,%s,&,==", ARG(1), ARG(0));
-		break;
-	case ARM_INS_LDRD:
-		addr &= ~3LL;
-		if (MEMDISP(2) < 0) {
-			const char *pc = "$$";
-			if (REGBASE(2) == ARM_REG_PC) {
-				op->refptr = 4;
-				op->ptr = addr + pcdelta + MEMDISP(2);
-				rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",2,2,%s,%d,+,>>,<<,+,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-					(ut64)MEMDISP(2), pc, pcdelta, REG(0), REG(1));
-			} else {
-				int disp = MEMDISP(2);
-				// not refptr, because we can't grab the reg value statically op->refptr = 4;
-				if (disp < 0) {
-					rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",%s,-,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-						(ut64)-disp, MEMBASE(2), REG(0), REG(1));
-				} else {
-					rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",%s,+,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-						(ut64)disp, MEMBASE(2), REG(0), REG(1));
-				}
-			}
-		} else {
-			if (REGBASE(2) == ARM_REG_PC) {
-				const char *pc = "$$";
-				op->refptr = 4;
-				op->ptr = addr + pcdelta + MEMDISP(2);
-				if (HASMEMINDEX(2) || ISREG(2)) {
-					const char op_index = ISMEMINDEXSUB(2) ? '-' : '+';
-					rz_strbuf_appendf(&op->esil, "%s,2,2,%d,%s,+,>>,<<,%c,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-						MEMINDEX(2), pcdelta, pc, op_index, REG(0), REG(1));
-				} else {
-					rz_strbuf_appendf(&op->esil, "2,2,%d,%s,+,>>,<<,%d,+,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-						pcdelta, pc, MEMDISP(2), REG(0), REG(1));
-				}
-			} else {
-				if (HASMEMINDEX(2)) { // e.g. `ldrd r2, r3 [r4, r1]`
-					const char op_index = ISMEMINDEXSUB(2) ? '-' : '+';
-					rz_strbuf_appendf(&op->esil, "%s,%s,%c,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-						MEMINDEX(2), MEMBASE(2), op_index, REG(0), REG(1));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%d,%s,+,0xffffffff,&,DUP,[4],%s,=,4,+,[4],%s,=",
-						MEMDISP(2), MEMBASE(2), REG(0), REG(1));
-				}
-				if (insn->detail->arm.writeback) {
-					if (ISPOSTINDEX32()) {
-						if (ISIMM(3)) {
-							rz_strbuf_appendf(&op->esil, ",%s,%d,+,%s,=",
-								MEMBASE(2), IMM(3), MEMBASE(2));
-						} else {
-							const char op_index = ISMEMINDEXSUB(3) ? '-' : '+';
-							rz_strbuf_appendf(&op->esil, ",%s,%s,%c,%s,=",
-								REG(3), MEMBASE(2), op_index, MEMBASE(2));
-						}
-					} else if (ISPREINDEX32()) {
-						if (HASMEMINDEX(2)) {
-							const char op_index = ISMEMINDEXSUB(2) ? '-' : '+';
-							rz_strbuf_appendf(&op->esil, ",%s,%s,%c,%s,=",
-								MEMINDEX(2), MEMBASE(2), op_index, MEMBASE(2));
-						} else {
-							rz_strbuf_appendf(&op->esil, ",%s,%d,+,%s,=",
-								MEMBASE(2), MEMDISP(2), MEMBASE(2));
-						}
-					}
-				}
-			}
-		}
-		break;
-	case ARM_INS_LDRB:
-		if (ISMEM(1) && LSHIFT2(1)) {
-			rz_strbuf_appendf(&op->esil, "%s,%d,%s,<<,+,0xffffffff,&,[1],0x%x,&,%s,=",
-				MEMBASE(1), LSHIFT2(1), MEMINDEX(1), mask, REG(0));
-		} else if (HASMEMINDEX(1)) {
-			rz_strbuf_appendf(&op->esil, "%s,%s,+,0xffffffff,&,[1],%s,=",
-				MEMINDEX(1), MEMBASE(1), REG(0));
-		} else {
-			rz_strbuf_appendf(&op->esil, "%s,%d,+,[1],%s,=",
-				MEMBASE(1), MEMDISP(1), REG(0));
-		}
-		if (insn->detail->arm.writeback) {
-			if (ISIMM(2)) {
-				rz_strbuf_appendf(&op->esil, ",%s,%d,+,%s,=",
-					MEMBASE(1), IMM(2), MEMBASE(1));
-			} else {
-				rz_strbuf_appendf(&op->esil, ",%s,%d,+,%s,=",
-					MEMBASE(1), MEMDISP(1), MEMBASE(1));
-			}
-		}
-		break;
-	case ARM_INS_SXTH:
-		rz_strbuf_appendf(&op->esil,
-			"15,%s,>>,1,&,?{,15,-1,<<,%s,0xffff,&,|,%s,:=,}{,%s,0xffff,%s,:=,}",
-			REG(1), REG(1), REG(0), REG(1), REG(0));
-		break;
-	case ARM_INS_SXTB:
-		rz_strbuf_appendf(&op->esil,
-			"7,%s,>>,1,&,?{,7,-1,<<,%s,0xff,&,|,%s,:=,}{,%s,0xff,&,%s,:=,}",
-			REG(1), REG(1), REG(0), REG(1), REG(0));
-		break;
-	case ARM_INS_LDREX:
-	case ARM_INS_LDREXB:
-	case ARM_INS_LDREXD:
-	case ARM_INS_LDREXH:
-		op->family = RZ_ANALYSIS_OP_FAMILY_THREAD;
-		// intentional fallthrough
-	case ARM_INS_LDRHT:
-	case ARM_INS_LDRH:
-	case ARM_INS_LDRT:
-	case ARM_INS_LDRBT:
-	case ARM_INS_LDRSB:
-	case ARM_INS_LDRSBT:
-	case ARM_INS_LDRSH:
-	case ARM_INS_LDRSHT:
-	case ARM_INS_LDR:
-		switch (insn->id) {
-		case ARM_INS_LDRHT:
-		case ARM_INS_LDRH:
-		case ARM_INS_LDRSH:
-		case ARM_INS_LDRSHT:
-			mask = UT16_MAX;
-			break;
-		default:
-			mask = UT32_MAX;
-			break;
-		}
-		addr &= ~3LL;
-		if (MEMDISP(1) < 0) {
-			const char *pc = "$$";
-			if (REGBASE(1) == ARM_REG_PC) {
-				op->refptr = 4;
-				op->ptr = addr + pcdelta + MEMDISP(1);
-				rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",2,2,%s,>>,<<,+,0xffffffff,&,[4],0x%x,&,%s,=",
-					(ut64)MEMDISP(1), pc, mask, REG(0));
-			} else {
-				int disp = MEMDISP(1);
-				// not refptr, because we can't grab the reg value statically op->refptr = 4;
-				if (disp < 0) {
-					rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",%s,-,0xffffffff,&,[4],0x%x,&,%s,=",
-						(ut64)-disp, MEMBASE(1), mask, REG(0));
-				} else {
-					rz_strbuf_appendf(&op->esil, "0x%" PFMT64x ",%s,+,0xffffffff,&,[4],0x%x,&,%s,=",
-						(ut64)disp, MEMBASE(1), mask, REG(0));
-				}
-			}
-		} else {
-			if (REGBASE(1) == ARM_REG_PC) {
-				const char *pc = "$$";
-				op->refptr = 4;
-				op->ptr = addr + pcdelta + MEMDISP(1);
-				if (ISMEM(1) && LSHIFT2(1)) {
-					rz_strbuf_appendf(&op->esil, "2,2,%d,%s,+,>>,<<,%d,%s,<<,+,0xffffffff,&,[4],0x%x,&,%s,=",
-						pcdelta, pc, LSHIFT2(1), MEMINDEX(1), mask, REG(0));
-				} else {
-					if (ISREG(1)) {
-						rz_strbuf_appendf(&op->esil, "2,2,%d,%s,+,>>,<<,%s,+,0xffffffff,&,[4],0x%x,&,%s,=",
-							pcdelta, pc, MEMINDEX(1), mask, REG(0));
-					} else {
-						rz_strbuf_appendf(&op->esil, "2,2,%d,%s,+,>>,<<,%d,+,0xffffffff,&,[4],0x%x,&,%s,=",
-							pcdelta, pc, MEMDISP(1), mask, REG(0));
-					}
-				}
-			} else {
-				if (ISMEM(1) && LSHIFT2(1)) {
-					rz_strbuf_appendf(&op->esil, "%s,%d,%s,<<,+,0xffffffff,&,[4],0x%x,&,%s,=",
-						MEMBASE(1), LSHIFT2(1), MEMINDEX(1), mask, REG(0));
-				} else if (HASMEMINDEX(1)) { // e.g. `ldr r2, [r3, r1]`
-					rz_strbuf_appendf(&op->esil, "%s,%s,+,0xffffffff,&,[4],0x%x,&,%s,=",
-						MEMINDEX(1), MEMBASE(1), mask, REG(0));
-				} else {
-					rz_strbuf_appendf(&op->esil, "%d,%s,+,0xffffffff,&,[4],0x%x,&,%s,=",
-						MEMDISP(1), MEMBASE(1), mask, REG(0));
-				}
-				if (insn->detail->arm.writeback) {
-					if (ISIMM(2)) {
-						rz_strbuf_appendf(&op->esil, ",%s,%d,+,%s,=",
-							MEMBASE(1), IMM(2), MEMBASE(1));
-					} else {
-						rz_strbuf_appendf(&op->esil, ",%s,%d,+,%s,=",
-							MEMBASE(1), MEMDISP(1), MEMBASE(1));
-					}
-				}
-			}
-		}
-		break;
-	case ARM_INS_MRS:
-		// TODO: esil for MRS
-		break;
-	case ARM_INS_MSR:
-		msr_flags = insn->detail->arm.operands[0].reg >> 4;
-		rz_strbuf_appendf(&op->esil, "0,");
-		if (msr_flags & 1) {
-			rz_strbuf_appendf(&op->esil, "0xFF,|,");
-		}
-		if (msr_flags & 2) {
-			rz_strbuf_appendf(&op->esil, "0xFF00,|,");
-		}
-		if (msr_flags & 4) {
-			rz_strbuf_appendf(&op->esil, "0xFF0000,|,");
-		}
-		if (msr_flags & 8) {
-			rz_strbuf_appendf(&op->esil, "0xFF000000,|,");
-		}
-		rz_strbuf_appendf(&op->esil, "DUP,!,SWAP,&,%s,SWAP,cpsr,&,|,cpsr,=", REG(1));
-		break;
-	case ARM_INS_UBFX:
-		if (IMM(3) > 0 && IMM(3) <= 32 - IMM(2)) {
-			rz_strbuf_appendf(&op->esil, "%d,%s,%d,%" PFMT64u ",<<,&,>>,%s,=",
-				IMM(2), REG(1), IMM(2), (ut64)bitmask_by_width[IMM(3) - 1], REG(0));
-		}
-		break;
-	case ARM_INS_UXTB:
-		rz_strbuf_appendf(&op->esil, "%s,0xff,&,%s,=", ARG(1), REG(0));
-		break;
-	case ARM_INS_RSB:
-		if (OPCOUNT() == 2) {
-			rz_strbuf_appendf(&op->esil, "%s,%s,-=", ARG(0), ARG(1));
-		} else if (OPCOUNT() == 3) {
-			rz_strbuf_appendf(&op->esil, "%s,%s,-,%s,=", ARG(1), ARG(2), ARG(0));
-		}
-		break;
-	case ARM_INS_BIC:
-		if (OPCOUNT() == 2) {
-			rz_strbuf_appendf(&op->esil, "%s,0xffffffff,^,%s,&=", ARG(1), ARG(0));
-		} else {
-			rz_strbuf_appendf(&op->esil, "%s,0xffffffff,^,%s,&,%s,=", ARG(2), ARG(1), ARG(0));
-		}
-		break;
-	case ARM_INS_SMMLA:
-		rz_strbuf_appendf(&op->esil, "32,%s,%s,*,>>,%s,+,0xffffffff,&,%s,=",
-			REG(1), REG(2), REG(3), REG(0));
-		break;
-	case ARM_INS_SMMLAR:
-		rz_strbuf_appendf(&op->esil, "32,0x80000000,%s,%s,*,+,>>,%s,+,0xffffffff,&,%s,=",
-			REG(1), REG(2), REG(3), REG(0));
-		break;
-	case ARM_INS_UMULL:
-		rz_strbuf_appendf(&op->esil, "32,%s,%s,*,DUP,0xffffffff,&,%s,=,>>,%s,=",
-			REG(2), REG(3), REG(0), REG(1));
-		break;
-	case ARM_INS_MLS:
-		rz_strbuf_appendf(&op->esil, "%s,%s,*,%s,-,0xffffffff,&,%s,=",
-			REG(1), REG(2), REG(3), REG(0));
-		break;
-	case ARM_INS_MLA:
-		rz_strbuf_appendf(&op->esil, "%s,%s,*,%s,+,0xffffffff,&,%s,=",
-			REG(1), REG(2), REG(3), REG(0));
-		break;
-	case ARM_INS_MVN:
-		rz_strbuf_appendf(&op->esil, "-1,%s,^,0xffffffff,&,%s,=",
-			ARG(1), REG(0));
-		break;
-	case ARM_INS_BFI: {
-		if (OPCOUNT() >= 3 && ISIMM(3) && IMM(3) > 0 && IMM(3) < 64) {
-			ut64 mask = bitmask_by_width[IMM(3) - 1];
-			ut64 shift = IMM(2);
-			ut64 notmask = ~(mask << shift);
-			// notmask,dst,&,lsb,mask,src,&,<<,|,dst,=
-			rz_strbuf_setf(&op->esil, "%" PFMT64u ",%s,&,%" PFMT64u ",%" PFMT64u ",%s,&,<<,|,0xffffffff,&,%s,=",
-				notmask, REG(0), shift, mask, REG(1), REG(0));
-		}
-		break;
-	}
-	case ARM_INS_BFC: {
-		if (OPCOUNT() >= 2 && ISIMM(2) && IMM(2) > 0 && IMM(2) < 64) {
-			ut64 mask = bitmask_by_width[IMM(2) - 1];
-			ut64 shift = IMM(1);
-			ut64 notmask = ~(mask << shift);
-			// notmask,dst,&,dst,=
-			rz_strbuf_setf(&op->esil, "%" PFMT64u ",%s,&,0xffffffff,&,%s,=",
-				notmask, REG(0), REG(0));
-		}
-		break;
-	}
-	case ARM_INS_REV: {
-		const char *r0 = REG(0);
-		const char *r1 = REG(1);
-		rz_strbuf_setf(&op->esil,
-			"24,0xff,%s,&,<<,%s,=,"
-			"16,0xff,8,%s,>>,&,<<,%s,|=,"
-			"8,0xff,16,%s,>>,&,<<,%s,|=,"
-			"0xff,24,%s,>>,&,%s,|=,",
-			r1, r0, r1, r0, r1, r0, r1, r0);
-		break;
-	}
-	case ARM_INS_REV16: {
-		const char *r0 = REG(0);
-		const char *r1 = REG(1);
-		rz_strbuf_setf(&op->esil,
-			"8,0xff00ff00,%s,&,>>,%s,=,"
-			"8,0x00ff00ff,%s,&,<<,%s,|=,",
-			r1, r0, r1, r0);
-		break;
-	}
-	case ARM_INS_REVSH: {
-		const char *r0 = REG(0);
-		const char *r1 = REG(1);
-		rz_strbuf_setf(&op->esil,
-			"8,0xff00,%s,&,>>,%s,=,"
-			"8,0x00ff,%s,&,<<,%s,|=,"
-			"0x8000,%s,&,?{,"
-			"0xffff0000,%s,|=,"
-			"}",
-			r1, r0, r1, r0, r0, r0);
-		break;
-	}
-	case ARM_INS_TBB:
-		rz_strbuf_appendf(&op->esil, "%s,%s,+,0xffffffff,&,DUP,[1],1,SWAP,<<,+,pc,+=",
-			MEMBASE(0), MEMINDEX(0));
-		break;
-	case ARM_INS_TBH:
-		rz_strbuf_appendf(&op->esil, "%s,%d,%s,<<,+,0xffffffff,&,[2],1,SWAP,<<,pc,+=",
-			MEMBASE(0), LSHIFT2(0), MEMINDEX(0));
-		break;
-	default:
-		break;
-	}
-	// Update flags if required...TODO different instructions update different flags, but this should fix
-	// many errors
-	if (insn->detail->arm.update_flags) {
-		switch (insn->id) {
-		case ARM_INS_CMP:
-			rz_strbuf_appendf(&op->esil, ",$z,zf,:=,31,$s,nf,:=,32,$b,!,cf,:=,31,$o,vf,:=");
-			break;
-		case ARM_INS_ADD:
-		case ARM_INS_RSB:
-		case ARM_INS_SUB:
-		case ARM_INS_SBC:
-		case ARM_INS_ADC:
-		case ARM_INS_CMN:
-			rz_strbuf_appendf(&op->esil, ",$z,zf,:=,31,$s,nf,:=,31,$c,cf,:=,31,$o,vf,:=");
-			break;
-		default:
-			rz_strbuf_appendf(&op->esil, ",$z,zf,:=,31,$s,nf,:=");
-		}
-	}
-
-	rz_strbuf_append(&op->esil, postfix);
-
-	return 0;
 }
 
 static int cond_cs2r2(int cc) {
@@ -3874,7 +1740,7 @@ static void op_fillval(RzAnalysis *analysis, RzAnalysisOp *op, csh handle, cs_in
 	}
 }
 
-static int analop(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, RzAnalysisOpMask mask) {
+static int analysis_op(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, int len, RzAnalysisOpMask mask) {
 	ArmCSContext *ctx = (ArmCSContext *)a->plugin_data;
 
 	cs_insn *insn = NULL;
@@ -3929,7 +1795,7 @@ static int analop(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, in
 				opex64(&op->opex, ctx->handle, insn);
 			}
 			if (mask & RZ_ANALYSIS_OP_MASK_ESIL) {
-				analop64_esil(a, op, addr, buf, len, &ctx->handle, insn);
+				rz_arm_cs_analysis_op_64_esil(a, op, addr, buf, len, &ctx->handle, insn);
 			}
 		} else {
 			anop32(a, ctx->handle, op, insn, thumb, (ut8 *)buf, len);
@@ -3937,7 +1803,7 @@ static int analop(RzAnalysis *a, RzAnalysisOp *op, ut64 addr, const ut8 *buf, in
 				opex(&op->opex, ctx->handle, insn);
 			}
 			if (mask & RZ_ANALYSIS_OP_MASK_ESIL) {
-				analop_esil(a, op, addr, buf, len, &ctx->handle, insn, thumb);
+				rz_arm_cs_analysis_op_32_esil(a, op, addr, buf, len, &ctx->handle, insn, thumb);
 			}
 		}
 		set_opdir(op);
@@ -4383,7 +2249,7 @@ static ut8 *analysis_mask(RzAnalysis *analysis, int size, const ut8 *data, ut64 
 			free(hint);
 		}
 
-		if ((oplen = analop(analysis, op, at + idx, data + idx, size - idx, RZ_ANALYSIS_OP_MASK_BASIC)) < 1) {
+		if ((oplen = analysis_op(analysis, op, at + idx, data + idx, size - idx, RZ_ANALYSIS_OP_MASK_BASIC)) < 1) {
 			break;
 		}
 		if (op->ptr != UT64_MAX || op->jump != UT64_MAX) {
@@ -4548,7 +2414,7 @@ RzAnalysisPlugin rz_analysis_plugin_arm_cs = {
 	.preludes = analysis_preludes,
 	.bits = 16 | 32 | 64,
 	.address_bits = address_bits,
-	.op = &analop,
+	.op = &analysis_op,
 	.init = &init,
 	.fini = &fini,
 };
